@@ -13,14 +14,24 @@ if(NOT VERSION AND GIT_FOUND)
                   OUTPUT_VARIABLE VERSION
                   OUTPUT_STRIP_TRAILING_WHITESPACE)
 endif()
-if(NOT VERSION MATCHES "^[0-9]+\.[0-9]+\.[0-9]+")
-    message(WARNING "Could not get version from git repository (defaulting to VERSION=0.0.0).")
-    set(VERSION "0.0.0")
-endif()
 configure_file("src/common/version.h.in" "${CMAKE_SOURCE_DIR}/src/common/_version.h")
 
-set(CMAKE_CXX_STANDARD 17)
+if(CMAKE_SYSTEM_NAME MATCHES "Darwin")
+  # Karabiner-DriverKit requires C++20
+  set(CMAKE_CXX_STANDARD 20)
+else()
+  set(CMAKE_CXX_STANDARD 17)
+endif()
 set(CMAKE_CXX_EXTENSIONS OFF)
+
+if(MSVC)
+  add_compile_options(/GR- /GL /Os /W1 /MP /W4 /permissive- /wd4100)
+  add_link_options(/LTCG)
+  add_compile_options($<$<CONFIG:RELEASE>:/MT>)
+  add_compile_definitions(_CRT_SECURE_NO_WARNINGS _WINSOCK_DEPRECATED_NO_WARNINGS)
+else()
+  add_compile_options(-Wall -Wpedantic -Wno-unknown-pragmas)
+endif()
 
 include_directories(src)
 
