@@ -1,10 +1,12 @@
 #pragma once
 
 #include "get_key_name.h"
+#include "StringTyper.h"
 #include "runtime/KeyEvent.h"
 #include <stdexcept>
 #include <string>
 #include <functional>
+#include <optional>
 
 // Here are some examples for input and output expressions. Each example
 // consists of a description of the desired result, followed by the
@@ -36,37 +38,40 @@
 //   Press C while holding A and B:
 //     (A B){C}   -> +A +B +C
 
-struct ParseError : std::runtime_error {
-  using std::runtime_error::runtime_error;
-};
-
 class ParseKeySequence {
 public:
+  struct ParseError : std::runtime_error {
+    using std::runtime_error::runtime_error;
+  };
   using GetKeyByName = std::function<Key(std::string_view)>;
   using AddTerminalCommand = std::function<Key(std::string_view)>;
 
-  KeySequence operator()(const std::string& str, bool is_input,
+  KeySequence operator()(std::string_view str, bool is_input,
     GetKeyByName get_key_by_name = ::get_key_by_name,
     AddTerminalCommand add_terminal_command = { });
 
 private:
-  using It = std::string::const_iterator;
+  using It = std::string_view::const_iterator;
 
   void parse(It it, const It end);
   Key read_key(It* it, const It end);
   void add_key_to_sequence(Key key, KeyState state);
   void add_key_to_buffer(Key key);
-  void add_timeout_event(KeyState state, uint16_t timeout);
+  void add_timeout_event(uint16_t timeout, bool is_not, bool cancel_on_up);
+  bool add_string_typing(std::string_view string);
   bool remove_from_keys_not_up(Key key);
   void flush_key_buffer(bool up_immediately);
   void up_any_keys_not_up_yet();
-  void sync_adjacent_to_timeout();
+  void sync_after_not_timeouts();
   bool all_pressed_at_once() const;
   void remove_any_up_from_end();
+  void check_ContextActive_usage();
 
+  std::string_view m_string;
   bool m_is_input{ };
   GetKeyByName m_get_key_by_name;
   AddTerminalCommand m_add_terminal_command;
+  std::optional<StringTyper> m_string_typer;
   std::vector<Key> m_keys_not_up;
   std::vector<Key> m_key_buffer;
   KeySequence m_sequence;
