@@ -1,25 +1,30 @@
 #pragma once
 
+#include "common/Host.h"
+#include "common/MessageType.h"
+#include "config/Config.h"
 #include <memory>
-#include <vector>
-#include "common/Connection.h"
-
-struct Config;
 
 class ServerPort {
-private:
-  std::unique_ptr<Connection> m_connection;
-
 public:
   ServerPort();
-  ServerPort(const ServerPort&) = delete;
-  ServerPort& operator=(const ServerPort&) = delete;
-  ~ServerPort();
-
-  Connection::Socket socket() const;
-  bool initialize();
+  Socket socket() const { return m_connection.socket(); }
+  bool connect();
+  void disconnect();
   bool send_config(const Config& config);
   bool send_active_contexts(const std::vector<int>& indices);
   bool send_validate_state();
-  bool receive_triggered_action(Duration timeout, int* triggered_action);
+  bool send_set_virtual_key_state(Key key, KeyState state);
+  bool send_request_device_names();
+
+  struct MessageHandler {
+    virtual void on_execute_action_message(int action_index) = 0;
+    virtual void on_virtual_key_state_message(Key key, KeyState state) = 0;
+    virtual void on_device_names_message(std::vector<std::string> devices) = 0;
+  };
+  bool read_messages(MessageHandler& handler, std::optional<Duration> timeout);
+
+private:
+  Host m_host;
+  Connection m_connection;
 };
