@@ -37,10 +37,19 @@ bool interpret_commandline(Settings& settings, int argc, char* argv[]) {
     else if (argument == T("--no-tray")) {
       settings.no_tray_icon = true;
     }
+    else if (argument == T("--no-notify")) {
+      settings.no_notify = true;
+    }
     else {
       return false;
     }
   }
+
+#if defined(__APPLE__)
+  // not implemented yet
+  settings.no_tray_icon = true;
+#endif
+
   return true;
 }
 
@@ -52,8 +61,9 @@ void print_help_message() {
     "  -c, --config <path>  configuration file.\n"
     "  -u, --update         reload configuration file when it changes.\n"
     "  -v, --verbose        enable verbose output.\n"
-    "  --no-tray            do not show tray icon/notifications.\n"
-    "  --check              check the config for errors.\n"
+    "  --no-notify          do not show notifications.\n"
+    "  --no-tray            do not show tray icon.\n"
+    "  --check              check the config for errors and exit.\n"
     "  -h, --help           print this help.\n"
     "\n"
     "%s\n"
