@@ -1,23 +1,9 @@
 
 #include "client/FocusedWindow.h"
 #include "common/windows/win.h"
+#include <psapi.h>
 #include <array>
 #include <cstring>
-
-namespace {
-  std::string wide_to_utf8(const std::wstring_view& str) {
-    auto result = std::string();
-    result.resize(WideCharToMultiByte(CP_UTF8, 0, 
-      str.data(), static_cast<int>(str.size()), 
-      NULL, 0, 
-      NULL, 0));
-    WideCharToMultiByte(CP_UTF8, 0, 
-      str.data(), static_cast<int>(str.size()), 
-      result.data(), static_cast<int>(result.size()),
-      NULL, 0);
-    return result;
-  }
-} // namespace
 
 class FocusedWindowImpl {
 private:
@@ -25,11 +11,13 @@ private:
   std::wstring m_current_title;
   std::string m_class;
   std::string m_title;
+  std::string m_path;
 
 public:
   HWND current() const { return m_current_window; }
   const std::string& window_class() const { return m_class; }
   const std::string& window_title() const { return m_title; }
+  const std::string& window_path() const { return m_path; }
 
   bool update() {
     const auto hwnd = GetForegroundWindow();
@@ -50,6 +38,19 @@ public:
     GetClassNameW(hwnd, buffer.data(), static_cast<int>(buffer.size()));
     m_class = wide_to_utf8(buffer.data());
     m_title = wide_to_utf8(m_current_title);
+
+#if (PSAPI_VERSION >= 2)
+    m_path.clear();
+    auto pid = DWORD{ };
+    auto buffer_size = DWORD{ buffer.size() };
+    if (GetWindowThreadProcessId(hwnd, &pid))
+      if (const auto handle = OpenProcess(PROCESS_QUERY_INFORMATION | PROCESS_VM_READ, FALSE, pid)) {
+        if (QueryFullProcessImageNameW(handle, 0, buffer.data(), &buffer_size))
+          m_path = wide_to_utf8(buffer.data());
+        CloseHandle(handle);
+      }
+#endif
+
     return true;
   }
 };
@@ -62,6 +63,13 @@ FocusedWindow::FocusedWindow(FocusedWindow&& rhs) noexcept = default;
 FocusedWindow& FocusedWindow::operator=(FocusedWindow&& rhs) noexcept = default;
 FocusedWindow::~FocusedWindow() = default;
 
+bool FocusedWindow::initialize() {
+  return true;
+}
+
+void FocusedWindow::shutdown() {
+}
+
 bool FocusedWindow::update() {
   return m_impl->update();
 }
@@ -72,6 +80,10 @@ const std::string& FocusedWindow::window_class() const {
 
 const std::string& FocusedWindow::window_title() const {
   return m_impl->window_title();
+}
+
+const std::string& FocusedWindow::window_path() const {
+  return m_impl->window_path();
 }
 
 bool FocusedWindow::is_inaccessible() const {
