@@ -1,8 +1,6 @@
 
-#include "client/FocusedWindow.h"
 #include "client/Settings.h"
-#include "client/ConfigFile.h"
-#include "client/ServerPort.h"
+#include "client/ClientState.h"
 #include "common/windows/LimitSingleInstance.h"
 #include "common/output.h"
 #include "Wtsapi32.h"
@@ -20,6 +18,7 @@ namespace {
   const auto WM_APP_RESET = WM_APP + 0;
   const auto WM_APP_SERVER_MESSAGE = WM_APP + 1;
   const auto WM_APP_TRAY_NOTIFY = WM_APP + 2;
+  const auto WM_APP_CONTROL_MESSAGE = WM_APP + 3;
   const auto TIMER_UPDATE_CONFIG = 1;
   const auto TIMER_UPDATE_CONTEXT = 2;
   const auto TIMER_CREATE_TRAY_ICON = 3;
@@ -29,149 +28,65 @@ namespace {
   const auto IDI_OPEN_CONFIG = 4;
   const auto IDI_UPDATE_CONFIG = 5;
   const auto IDI_ABOUT = 6;
+  const auto IDI_DEVICES = 7;
 
   const auto update_context_inverval_ms = 50;
   const auto update_config_interval_ms = 500;
   const auto recreate_tray_icon_interval_ms = 1000;
 
   Settings g_settings;
-  ConfigFile g_config_file;
-  ServerPort g_server;
-  FocusedWindow g_focused_window;
-  std::vector<int> g_new_active_contexts;
-  std::vector<int> g_current_active_contexts;
+  ClientState g_state;
   bool g_was_inaccessible;
   bool g_session_changed;
   HWND g_window;
   NOTIFYICONDATAW g_tray_icon;
-  bool g_active{ true };
-  
-  std::wstring utf8_to_wide(std::string_view str) {
-    auto result = std::wstring();
-    result.resize(MultiByteToWideChar(CP_UTF8, 0, 
-      str.data(), static_cast<int>(str.size()), 
-      NULL, 0));
-    MultiByteToWideChar(CP_UTF8, 0, 
-      str.data(), static_cast<int>(str.size()), 
-      result.data(), static_cast<int>(result.size()));
-    return result;
-  }
 
-  bool starts_with_case_insensitive(const std::string& string, const char* value) {
-    for (const auto* a = string.c_str(), *b = value; *b != '\0'; ++a, ++b)
-      if (*a == '\0' || 
-          std::tolower(static_cast<int>(*a)) != std::tolower(static_cast<int>(*b)))
-        return false;
-    return true;
-  }
-
-  bool execute_terminal_command(const std::string& command) {
-    SetForegroundWindow(g_window);
-
-    auto cmd = std::wstring(MAX_PATH, ' ');
-    cmd.resize(GetSystemDirectoryW(cmd.data(), static_cast<UINT>(cmd.size())));
-    cmd += L"\\CMD.EXE";
-
-    auto args = L"/C " + utf8_to_wide(command);
-
-    auto flags = DWORD{ };
-    if (!starts_with_case_insensitive(command, "cmd") && 
-        !starts_with_case_insensitive(command, "powershell"))
-      flags |= CREATE_NO_WINDOW;
-
-    auto startup_info = STARTUPINFOW{ sizeof(STARTUPINFOW) };
-    auto process_info = PROCESS_INFORMATION{ };
-    if (!CreateProcessW(cmd.data(), args.data(), nullptr, nullptr, FALSE, 
-        flags, nullptr, nullptr, &startup_info, &process_info)) 
-      return false;
-    
-    CloseHandle(process_info.hProcess);
-    CloseHandle(process_info.hThread);
-    return true;
-  }
-
-  void execute_action(int triggered_action) {
-    const auto& actions = g_config_file.config().actions;
-    if (triggered_action >= 0 &&
-        triggered_action < static_cast<int>(actions.size())) {
-      const auto& action = actions[triggered_action];
-      const auto& command = action.terminal_command;
-      verbose("Executing terminal command '%s'", command.c_str());
-      execute_terminal_command(command);
-    }
-  }
- 
-  void update_active_contexts(bool force_send) {
-    const auto& contexts = g_config_file.config().contexts;
-
-    g_new_active_contexts.clear();
-    if (g_active)
-      for (auto i = 0; i < static_cast<int>(contexts.size()); ++i)
-        if (contexts[i].matches(g_focused_window.window_class(),
-            g_focused_window.window_title()))
-          g_new_active_contexts.push_back(i);
-
-    if (force_send || g_new_active_contexts != g_current_active_contexts) {
-      verbose("Active contexts updated (%u)", g_new_active_contexts.size());
-      g_server.send_active_contexts(g_new_active_contexts);
-      g_current_active_contexts.swap(g_new_active_contexts);
-    }
-  }
-  
   void validate_state() {
     // validate internal state when a window of another user was focused
     // force validation after session change
     const auto check_accessibility = 
       !std::exchange(g_session_changed, false);
     if (check_accessibility) {
-      if (g_focused_window.is_inaccessible()) {
+      if (g_state.is_focused_window_inaccessible()) {
         g_was_inaccessible = true;
         return;
       }
       if (!std::exchange(g_was_inaccessible, false))
         return;
     }
-    g_server.send_validate_state();
-  }
-
-  void update_context() {
-    if (g_focused_window.update()) {
-      verbose("Detected focused window changed:");
-      verbose("  class = '%s'", g_focused_window.window_class().c_str());
-      verbose("  title = '%s'", g_focused_window.window_title().c_str());
-      update_active_contexts(false);
-    }
-  }
-
-  bool send_config() {
-    if (!g_server.send_config(g_config_file.config())) {
-      error("Sending configuration failed");
-      return false;
-    }
-    update_active_contexts(true);
-    return true;
+    verbose("Validating state");
+    g_state.send_validate_state();
   }
 
   bool connect() {
-    verbose("Connecting to keymapperd");
-    if (!g_server.initialize() ||
-        WSAAsyncSelect(g_server.socket(), g_window,
-          WM_APP_SERVER_MESSAGE, (FD_READ | FD_CLOSE)) != 0) {
-      error("Connecting to keymapperd failed");
-      return false;
-    }
-    return true;
+    if (auto socket = g_state.connect_server(); 
+        socket && WSAAsyncSelect(*socket, g_window, 
+          WM_APP_SERVER_MESSAGE, (FD_READ | FD_CLOSE)) == 0)
+      return true;
+    error("Connecting to keymapperd failed");
+    return false;
   }
 
-  void update_config(bool check_modified = true) {
-    if (g_config_file.update(check_modified)) {
-      message("Configuration updated");
-      send_config();
-    }    
+  bool listen_for_control() {
+    if (auto socket = g_state.listen_for_control_connections();
+        socket && WSAAsyncSelect(*socket, g_window,
+          WM_APP_CONTROL_MESSAGE, FD_ACCEPT) == 0)
+      return true;
+    error("Initializing keymapperctl connection failed");
+    return false;
+  }
+
+  bool accept_control() {
+    if (auto socket = g_state.accept_control_connection();
+        socket && WSAAsyncSelect(*socket, g_window,
+          WM_APP_CONTROL_MESSAGE, (FD_READ | FD_CLOSE)) == 0)
+      return true;
+    error("Accepting keymapperctl connection failed");
+    return false;
   }
 
   void open_configuration() {
-    const auto filename = g_config_file.filename().wstring();
+    const auto filename = g_state.config_filename().wstring();
 
     // create if it does not exist
     if (auto handle = CreateFileW(filename.c_str(), GENERIC_READ, FILE_SHARE_READ, 
@@ -203,18 +118,14 @@ namespace {
     MessageBoxA(g_window, text.data(), caption, MB_ICONINFORMATION | MB_TOPMOST);
   }
 
-  void toggle_active() {
-    g_active = !g_active;
-    update_active_contexts(true);
-  }
-
   void open_tray_menu() {
     auto popup_menu = CreatePopupMenu();
     AppendMenuW(popup_menu, 
-      (g_active ? MF_CHECKED : MF_UNCHECKED) | MF_STRING, IDI_ACTIVE, L"Active");
+      (g_state.is_active() ? MF_CHECKED : MF_UNCHECKED) | MF_STRING, IDI_ACTIVE, L"Active");
     AppendMenuW(popup_menu, MF_STRING, IDI_OPEN_CONFIG, L"Configuration");
     if (!g_settings.auto_update_config)
       AppendMenuW(popup_menu, MF_STRING, IDI_UPDATE_CONFIG, L"Reload");
+    AppendMenuW(popup_menu, MF_STRING, IDI_DEVICES, L"Devices");
     AppendMenuW(popup_menu, MF_STRING, IDI_HELP, L"Help");
     AppendMenuW(popup_menu, MF_STRING, IDI_ABOUT, L"About");
     AppendMenuW(popup_menu, MF_SEPARATOR, 0, nullptr);
@@ -243,29 +154,36 @@ namespace {
         if (lparam == WM_LBUTTONUP || lparam == WM_RBUTTONUP)
           open_tray_menu();
         else if (lparam == WM_MBUTTONUP)
-          toggle_active();
+          g_state.toggle_active();
         else if (lparam == 0x405)
           open_configuration();
         return 0;
 
       case WM_APP_SERVER_MESSAGE:
         if (lparam == FD_READ) {
-          auto triggered_action = -1;
-          g_server.receive_triggered_action(Duration::zero(), &triggered_action);
-          execute_action(triggered_action);
+          g_state.read_server_messages();
         }
         else {
           verbose("Connection to keymapperd lost");
           verbose("---------------");
-          if (!connect() || !send_config())
+          if (!connect() || !g_state.send_config())
             PostQuitMessage(1);
+        }
+        return 0;
+
+      case WM_APP_CONTROL_MESSAGE:
+        if (lparam == FD_ACCEPT) {
+          accept_control();
+        }
+        else {
+          g_state.read_control_messages();
         }
         return 0;
 
       case WM_COMMAND:
         switch (wparam) {
           case IDI_ACTIVE:
-            toggle_active();
+            g_state.toggle_active();
             return 0;
 
           case IDI_OPEN_CONFIG:
@@ -273,7 +191,12 @@ namespace {
             return 0;
 
           case IDI_UPDATE_CONFIG:
-            update_config(false);
+            if (g_state.update_config(false))
+              g_state.send_config();
+            return 0;
+
+          case IDI_DEVICES:
+            g_state.request_device_names();
             return 0;
 
           case IDI_HELP:
@@ -292,11 +215,13 @@ namespace {
 
       case WM_TIMER: {
         if (wparam == TIMER_UPDATE_CONTEXT) {
-          update_context();
+          if (g_state.update_active_contexts())
+            g_state.send_active_contexts();
           validate_state();
         }
         else if (wparam == TIMER_UPDATE_CONFIG) {
-          update_config();
+          if (g_state.update_config(true))
+            g_state.send_config();
         }
         else if (wparam == TIMER_CREATE_TRAY_ICON) {
           // workaround: re/create tray icon after taskbar was created
@@ -354,14 +279,14 @@ int WINAPI wWinMain(HINSTANCE instance, HINSTANCE, LPWSTR, int) {
     return 1;
   }
   g_verbose_output = g_settings.verbose;
-  if (!g_settings.no_tray_icon)
+  if (!g_settings.no_notify)
     g_show_notification = &show_notification;
 
   g_settings.config_file_path = 
     resolve_config_file_path(g_settings.config_file_path);
 
   if (g_settings.check_config) {
-    if (!g_config_file.load(g_settings.config_file_path))
+    if (!g_state.load_config(g_settings.config_file_path))
       return 1;
     message("The configuration is valid");
     return 0;
@@ -373,6 +298,8 @@ int WINAPI wWinMain(HINSTANCE instance, HINSTANCE, LPWSTR, int) {
     error("Another instance is already running");
     return 1;
   }
+
+  SetPriorityClass(GetCurrentProcess(), HIGH_PRIORITY_CLASS);
 
   const auto window_class_name = L"Keymapper";
   auto window_class = WNDCLASSEXW{ };
@@ -390,6 +317,8 @@ int WINAPI wWinMain(HINSTANCE instance, HINSTANCE, LPWSTR, int) {
   if (!connect())
     return 1;
 
+  listen_for_control();
+
   if (!g_settings.no_tray_icon) {
     auto& icon = g_tray_icon;
     icon.cbSize = sizeof(icon);
@@ -406,9 +335,12 @@ int WINAPI wWinMain(HINSTANCE instance, HINSTANCE, LPWSTR, int) {
   }
 
   verbose("Loading configuration file '%ws'", g_settings.config_file_path.c_str());
-  g_config_file.load(g_settings.config_file_path);
-
-  send_config();
+  if (!g_state.load_config(g_settings.config_file_path)) {
+    // exit when there is no configuration and user cannot update it
+    if (g_settings.no_tray_icon)
+      return 1;
+  }
+  g_state.send_config();
   
   WTSRegisterSessionNotification(g_window, NOTIFY_FOR_THIS_SESSION);
 
@@ -419,7 +351,6 @@ int WINAPI wWinMain(HINSTANCE instance, HINSTANCE, LPWSTR, int) {
     SetTimer(g_window, TIMER_UPDATE_CONFIG, update_config_interval_ms, NULL);
   SetTimer(g_window, TIMER_UPDATE_CONTEXT, update_context_inverval_ms, NULL);
 
-  verbose("Entering update loop");
   auto message = MSG{ };
   while (GetMessageW(&message, nullptr, 0, 0) > 0) {
     TranslateMessage(&message);
