@@ -8,11 +8,16 @@
 bool g_verbose_output = false;
 void(*g_show_notification)(const char* message);
 
+constexpr const char* current_year() {
+  const auto date = __DATE__;
+  return &date[7];
+}
+
 const auto about_header_str = std::string(
 #if __has_include("_version.h")
 # include "_version.h"
 #endif
-  "\n(c) 2019-") + (__DATE__ + 7) + " by Albert Kalchmair";
+  "\n(c) 2019-") + current_year() + " by Albert Kalchmair";
 
 const char* about_header = about_header_str.c_str();
 const char* about_footer =
@@ -29,7 +34,7 @@ extern void show_notification(const char* message);
 
 namespace {
   void vprint(const char* format, va_list args, bool notify, bool is_error) {
-    auto buffer = std::array<char, 1024>();
+    auto buffer = std::array<char, 2048>();
     std::vsnprintf(buffer.data(), buffer.size(), format, args);
 
 #if defined(_WIN32)
