@@ -20,8 +20,13 @@ struct Config {
   struct Filter {
     std::string string;
     std::optional<std::regex> regex;
+    bool invert{ };
 
     bool matches(const std::string& text, bool substring) const {
+      return matches_uninverted(text, substring) ^ invert;
+    }
+
+    bool matches_uninverted(const std::string& text, bool substring) const {
       if (string.empty())
         return true;
       if (regex.has_value())
@@ -33,19 +38,27 @@ struct Config {
   };
 
   struct Context {
-    bool system_filter_matched;
+    bool system_filter_matched{ };
     Filter window_class_filter;
     Filter window_title_filter;
+    Filter window_path_filter;
     std::string device_filter;
+    KeySequence modifier_filter;
     std::vector<Input> inputs;
     std::vector<KeySequence> outputs;
     std::vector<CommandOutput> command_outputs;
+    bool invert_device_filter{ };
+    bool invert_modifier_filter{ };
+    bool fallthrough{ };
 
     bool matches(const std::string& window_class,
-                 const std::string& window_title) const {
+                 const std::string& window_title,
+                 const std::string& window_path) const {
       if (!window_class_filter.matches(window_class, false))
         return false;
       if (!window_title_filter.matches(window_title, true))
+        return false;
+      if (!window_path_filter.matches(window_path, true))
         return false;
       return true;
     }
@@ -57,4 +70,5 @@ struct Config {
 
   std::vector<Context> contexts;
   std::vector<Action> actions;
+  std::vector<std::pair<std::string, Key>> virtual_key_aliases;
 };
