@@ -16,13 +16,15 @@ namespace {
   const auto current_system = "linux";
 #elif defined(_WIN32)
   const auto current_system = "windows";
+#elif defined(__APPLE__)
+  const auto current_system = "macos";
 #else
 #  error unknown system
 #endif
 
   std::string to_lower(std::string str) {
     for (auto& c : str)
-      c = static_cast<char>(std::tolower(c));
+      c = static_cast<char>(std::tolower(static_cast<unsigned char>(c)));
     return str;
   }
 
@@ -50,15 +52,73 @@ namespace {
         ++it;
       }
   }
+
+  std::string_view trim(std::string_view s) {
+    while (!s.empty() && std::isspace(static_cast<unsigned char>(s.front())))
+      s.remove_prefix(1);
+    while (!s.empty() && std::isspace(static_cast<unsigned char>(s.back())))
+      s.remove_suffix(1);
+    return s;
+  }
+
+  std::string_view unquote(std::string_view s) {
+    if (s.size() >= 2 &&
+        s.front() == s.back() &&
+        (s.front() == '\'' || s.front() == '"'))
+      s = s.substr(1, s.size() - 2);
+    return s;
+  }
+
+  std::vector<std::string> get_argument_list(std::string_view list) {
+    auto it = list.begin();
+    const auto end = list.end();
+    auto begin = it;
+    auto arguments = std::vector<std::string>();
+    while (begin != end) {
+      skip_until_not_in_string(&it, end, ",");
+      if (it > begin) {
+        auto argument = std::string_view(
+          &*begin, std::distance(begin, it) - 1);
+        arguments.emplace_back(trim(argument));
+      }
+      begin = it;
+    }
+    return arguments;
+  }
+
+  std::string substitute_arguments(const std::string& text,
+      const std::vector<std::string>& arguments) {
+    auto it = text.begin();
+    const auto end = text.end();
+    auto result = std::string();
+    for (auto begin = it; it != end; begin = it) {
+      if (skip_until(&it, end, '$')) {
+        // argument or beginning of terminal command
+        result.append(begin, it - 1);
+        begin = it - 1;
+
+        if (auto number = try_read_number(&it, end)) {
+          const auto index = static_cast<size_t>(*number);
+          // ignore missing arguments
+          if (index < arguments.size())
+            result.append(unquote(arguments[index]));
+          continue;
+        }
+      }
+      result.append(begin, it);
+    }
+    return result;
+  }
 } // namespace
 
-Config ParseConfig::operator()(std::istream& is) {
+Config ParseConfig::operator()(std::istream& is) try {
   m_line_no = 0;
   m_config = { };
   m_commands.clear();
   m_macros.clear();
   m_logical_keys.clear();
-  m_context_modifier.clear();
+  m_system_filter_matched = true;
+  m_after_empty_context_block = false;
 
   // add default context
   m_config.contexts.push_back({ true, {}, {} });
@@ -78,39 +138,50 @@ Config ParseConfig::operator()(std::istream& is) {
   // check if there is a mapping for each command (to reduce typing errors)
   for (const auto& command : m_commands)
     if (!command.mapped)
-      throw ParseError("Command '" + command.name + "' was not mapped");
+      throw ConfigError("Command '" + command.name + "' was not mapped");
 
   // remove contexts of other systems or which are empty
-  m_config.contexts.erase(
-    std::remove_if(begin(m_config.contexts), end(m_config.contexts),
-      [](const Config::Context& context) {
-        return !context.system_filter_matched || 
-          (context.inputs.empty() && context.command_outputs.empty());
-      }),
-    m_config.contexts.end());
+  optimize_contexts();
 
   // replace logical keys (in reverse order of registration)
   for (auto it = m_logical_keys.rbegin(); it != m_logical_keys.rend(); ++it) {
     const auto& [name, both, left, right] = *it;
     replace_logical_key(both, left, right);
   }
+
+  // collect virtual key aliases
+  for (const auto& [name, value] : m_macros)
+    if (auto key = get_key_by_name(value); is_virtual_key(key))
+      m_config.virtual_key_aliases.emplace_back(name, key);
+
   return std::move(m_config);
 }
+catch (const ConfigError&) {
+  throw;
+}
+catch (const std::exception& ex) {
+  error(ex.what());
+}
 
-void ParseConfig::error(std::string message) {
-  throw ParseError(std::move(message) +
+void ParseConfig::error(std::string message) const {
+  throw ConfigError(std::move(message) +
     " in line " + std::to_string(m_line_no));
 }
 
 void ParseConfig::parse_line(It it, It end) {
   skip_space_and_comments(&it, end);
-  if (it == end)
-    return;
-
   if (skip(&it, end, "[")) {
+    if (m_after_empty_context_block)
+      m_config.contexts.back().fallthrough = true;
+
     parse_context(&it, end);
+    m_after_empty_context_block = true;
   }
   else {
+    m_after_empty_context_block = false;
+    if (it == end)
+      return;
+
     const auto begin = it;
     skip_ident(&it, end);
     auto first_ident = std::string(begin, it);
@@ -162,14 +233,21 @@ std::string ParseConfig::read_filter_string(It* it, const It end) {
   }
   else if (skip(it, end, "'") || skip(it, end, "\"")) {
     // a string
-    const char mark[2] = { *(*it - 1), '\0' };
-    if (!skip_until(it, end, mark))
+    if (!skip_until(it, end, *std::prev(*it)))
       error("Unterminated string");
     return std::string(begin + 1, *it - 1);
   }
   else {
     skip_value(it, end);
-    return std::string(begin, *it);
+    auto ident = preprocess_ident(std::string(begin, *it));
+
+    // trim quotes after macro substitution
+    if (!ident.empty() && (ident.front() == '\'' || ident.front() == '\"')) {
+      if (ident.size() < 2 || ident.back() != ident.front())
+        error("Unterminated string");
+      ident = ident.substr(1, ident.size() - 2);
+    }
+    return ident;
   }
 }
 
@@ -182,15 +260,31 @@ Config::Filter ParseConfig::read_filter(It* it, const It end) {
   return { std::move(string), { } };
 }
 
+KeySequence ParseConfig::parse_modifier_list(std::string_view string) {
+  auto it = string.begin();
+  const auto end = string.end();
+  auto sequence = KeySequence();
+  for (;;) {
+    skip_space(&it, end);
+    if (it == end)
+      break;
+    const auto is_not = skip(&it, end, "!");
+    const auto name = read_ident(&it, end);
+    const auto key = get_key_by_name(name);
+    if (!*key)
+      error("Invalid key name '" + name + "'");
+    sequence.emplace_back(key, (is_not ? KeyState::Not : KeyState::Down));
+  }
+  if (contains(sequence, Key::ContextActive))
+    error("Not allowed key ContextActive");
+  return sequence;
+}
+
 void ParseConfig::parse_context(It* it, const It end) {
+  auto& context = m_config.contexts.emplace_back();
+  context.system_filter_matched = true;
+
   skip_space(it, end);
-
-  auto system_filter_matched = true;
-  auto class_filter = Config::Filter();
-  auto title_filter = Config::Filter();
-  auto device_filter = std::string();
-  m_context_modifier = { };
-
   if (skip(it, end, "default")) {
     skip_space(it, end);
     if (!skip(it, end, "]"))
@@ -203,34 +297,37 @@ void ParseConfig::parse_context(It* it, const It end) {
         error("Identifier expected");
 
       skip_space(it, end);
+      const auto invert = skip(it, end, "!");
       if (!skip(it, end, "="))
         error("Missing '='");
 
       skip_space(it, end);
       if (attrib == "class") {
-        class_filter = read_filter(it, end);
+        context.window_class_filter = read_filter(it, end);
+        context.window_class_filter.invert = invert;
       }
       else if (attrib == "title") {
-        title_filter = read_filter(it, end);
+        context.window_title_filter = read_filter(it, end);
+        context.window_title_filter.invert = invert;
+      }
+      else if (attrib == "path") {
+        context.window_path_filter = read_filter(it, end);
+        context.window_path_filter.invert = invert;
       }
       else if (attrib == "system") {
-        system_filter_matched =
-          (to_lower(read_value(it, end)) == current_system);
+        const auto system = to_lower(read_value(it, end));
+        context.system_filter_matched = (system == current_system) ^ invert;
       }
       else if (attrib == "device") {
-        device_filter = read_filter(it, end).string;
-        if (device_filter.empty())
+        context.device_filter = read_filter(it, end).string;
+        if (context.device_filter.empty())
           error("String expected");
+        context.invert_device_filter = invert;
       }
       else if (attrib == "modifier") {
-        auto modifier = read_value(it, end);
-        m_context_modifier = m_parse_sequence(
-          preprocess(modifier.begin(), modifier.end()), true,
-          std::bind(&ParseConfig::get_key_by_name, this, _1));
-        m_context_modifier.erase(
-          std::remove_if(m_context_modifier.begin(), m_context_modifier.end(),
-            [](const KeyEvent& event) { return (event.state == KeyState::UpAsync); }),
-          m_context_modifier.end());
+        const auto modifier = preprocess(read_value(it, end));
+        context.modifier_filter = parse_modifier_list(modifier);
+        context.invert_modifier_filter = invert;
       }
       else {
         error("Unexpected '" + attrib + "'");
@@ -240,18 +337,15 @@ void ParseConfig::parse_context(It* it, const It end) {
       if (skip(it, end, "]"))
         break;
 
+      // allow to separate with commas
+      skip(it, end, ',');
+
       skip_space(it, end);
       if (*it == end)
         error("Missing ']'");
     }
   }
-
-  m_config.contexts.push_back({
-    system_filter_matched,
-    std::move(class_filter),
-    std::move(title_filter),
-    std::move(device_filter)
-  });
+  m_system_filter_matched = context.system_filter_matched;
 }
 
 void ParseConfig::parse_mapping(const std::string& name, It begin, It end) {
@@ -287,16 +381,10 @@ void ParseConfig::parse_command_and_mapping(const It in_begin, const It in_end,
     add_mapping(std::move(input), parse_output(out_begin, out_end));
 }
 
-KeySequence ParseConfig::parse_input(It it, It end) try {
+KeySequence ParseConfig::parse_input(It it, It end) {
   skip_space(&it, end);
-  auto sequence = m_parse_sequence(preprocess(it, end), true,
+  return m_parse_sequence(preprocess(it, end), true,
     std::bind(&ParseConfig::get_key_by_name, this, _1));
-  sequence.insert(sequence.begin(),
-    m_context_modifier.begin(), m_context_modifier.end());
-  return sequence;
-}
-catch (const std::exception& ex) {
-  error(ex.what());
 }
 
 Key ParseConfig::get_key_by_name(std::string_view name) const {
@@ -318,20 +406,21 @@ Key ParseConfig::add_terminal_command_action(std::string_view command) {
   return action_key_code;
 }
 
-KeySequence ParseConfig::parse_output(It it, It end) try {
+KeySequence ParseConfig::parse_output(It it, It end) {
   skip_space(&it, end);
-  return m_parse_sequence(preprocess(it, end), false,
+  auto sequence = m_parse_sequence(preprocess(it, end), false,
     std::bind(&ParseConfig::get_key_by_name, this, _1),
     std::bind(&ParseConfig::add_terminal_command_action, this, _1));
-}
-catch (const std::exception& ex) {
-  error(ex.what());
+  if (contains(sequence, Key::ContextActive))
+    error("Not allowed key ContextActive");
+  return sequence;
 }
 
 void ParseConfig::parse_macro(std::string name, It it, const It end) {
   if (*get_key_by_name(name))
     error("Invalid macro name '" + name + "'");
-  m_macros[std::move(name)] = preprocess(it, end);
+  if (m_system_filter_matched)
+    m_macros[std::move(name)] = preprocess(it, end);
 }
 
 bool ParseConfig::parse_logical_key_definition(
@@ -364,9 +453,26 @@ bool ParseConfig::parse_logical_key_definition(
 }
 
 std::string ParseConfig::preprocess_ident(std::string ident) const {
+  if (auto pos = ident.find('['); pos != std::string::npos) {
+    // macro with arguments
+    const auto name = ident.substr(0, pos);
+    const auto macro = m_macros.find(name);
+    if (macro == cend(m_macros))
+      error("Unknown macro '" + name + "'");
+
+    // preprocess arguments
+    auto arguments = get_argument_list(ident.substr(pos + 1));
+    for (auto& argument : arguments)
+      argument = preprocess(argument);
+
+    // substitute $0... in macro text with arguments
+    return substitute_arguments(macro->second, arguments);
+  }
+
   const auto macro = m_macros.find(ident);
   if (macro != cend(m_macros))
     return macro->second;
+
   return ident;
 }
 
@@ -381,17 +487,30 @@ std::string ParseConfig::preprocess(It it, const It end) const {
 
     // try to read ident
     auto begin = it;
-    skip_ident(&it, end);
+    if (!skip_ident_with_arglist(&it, end))
+      error("Incomplete argument list");
+
     if (begin != it) {
       // match read ident
       result.append(preprocess_ident(std::string(begin, it)));
     }
+    else if (*it == '\'' || *it == '\"') {
+      // a string
+      if (!skip_until(&it, end, *it++))
+        error("Unterminated string");
+
+      result.append(begin, it);
+    }
     else {
-      // output single character
+      // single character
       result.append(begin, ++it);
     }
   }
   return result;
+}
+
+std::string ParseConfig::preprocess(const std::string& string) const {
+  return preprocess(string.begin(), string.end());
 }
 
 Config::Context& ParseConfig::current_context() {
@@ -454,7 +573,8 @@ Key ParseConfig::add_logical_key(std::string name, Key left, Key right) {
 }
 
 void ParseConfig::replace_logical_key(Key both, Key left, Key right) {
-  for (auto& context : m_config.contexts) {
+  auto& contexts = m_config.contexts;
+  for (auto& context : contexts) {
     // replace !<both> with !<left> !<right>
     for (auto& input : context.inputs)
       replace_not_key(input.input, both, left, right);
@@ -489,5 +609,57 @@ void ParseConfig::replace_logical_key(Key both, Key left, Key right) {
       replace_key(output, both, left);
     for (auto& command : context.command_outputs)
       replace_key(command.output, both, left);
+
+    // replace !<both> with !<left> !<right> in context modifier
+    replace_not_key(context.modifier_filter, both, left, right);
+  }
+
+  // insert fallthrough context and replace the logical with a physical key
+  for (auto it = begin(contexts); it != end(contexts); ++it) {
+    if (contains(it->modifier_filter, both)) {
+      // duplicate and replace with <left> and <right>
+      it = contexts.insert(it, *it);
+      it->inputs.clear();
+      it->outputs.clear();
+      it->command_outputs.clear();
+      it->fallthrough = true;
+      replace_key(it->modifier_filter, both, left);
+      ++it;
+      replace_key(it->modifier_filter, both, right);
+    }
+  }
+}
+
+void ParseConfig::optimize_contexts() {
+  auto& contexts = m_config.contexts;
+  auto before_context = false;
+  for (auto i = static_cast<int>(contexts.size()) - 1; i >= 0; --i) {
+    const auto& context = contexts[i];
+    const auto can_not_match = !context.system_filter_matched;
+    if (context.fallthrough) {
+      // remove fallthrough contexts which are not before a context
+      if (can_not_match || !before_context)
+        contexts.erase(std::next(contexts.begin(), i));
+    }
+    else {
+      const auto has_no_effect = (context.inputs.empty() && context.command_outputs.empty());
+      if (can_not_match || has_no_effect) {
+        // convert fallthrough context when removing the non-fallthrough context
+        if (i > 0 && contexts[i - 1].fallthrough) {
+          auto& before = contexts[i - 1];
+          before.inputs = std::move(context.inputs);
+          before.outputs = std::move(context.outputs);
+          before.command_outputs = std::move(context.command_outputs);
+          before.fallthrough = false;
+        }
+        contexts.erase(std::next(contexts.begin(), i));
+
+        // no longer before a context when removing a non-fallthrough context
+        before_context = false;
+      }
+      else {
+        before_context = true;
+      }
+    }
   }
 }
