@@ -8,6 +8,10 @@
 
 class ParseConfig {
 public:
+  struct ConfigError : std::runtime_error {
+    using std::runtime_error::runtime_error;
+  };
+
   Config operator()(std::istream& is);
 
 private:
@@ -26,9 +30,10 @@ private:
 
   using It = std::string::const_iterator;
 
-  [[noreturn]] void error(std::string message);
+  [[noreturn]] void error(std::string message) const;
   void parse_line(It begin, It end);
   void parse_context(It* begin, It end);
+  KeySequence parse_modifier_list(std::string_view string);
   void parse_macro(std::string name, It begin, It end);
   bool parse_logical_key_definition(const std::string& name, It it, It end);
   void parse_mapping(const std::string& name, It begin, It end);
@@ -39,12 +44,14 @@ private:
   KeySequence parse_output(It begin, It end);
   std::string preprocess_ident(std::string ident) const;
   std::string preprocess(It begin, It end) const;
+  std::string preprocess(const std::string& string) const;
   Key add_logical_key(std::string name, Key left, Key right);
   void replace_logical_key(Key both, Key left, Key right);
   std::string read_filter_string(It* it, It end);
   Config::Filter read_filter(It* it, It end);
   Key get_key_by_name(std::string_view name) const;
   Key add_terminal_command_action(std::string_view command);
+  void optimize_contexts();
 
   Config::Context& current_context();
   Command* find_command(const std::string& name);
@@ -58,5 +65,6 @@ private:
   std::map<std::string, std::string> m_macros;
   std::vector<LogicalKey> m_logical_keys;
   ParseKeySequence m_parse_sequence;
-  KeySequence m_context_modifier;
+  bool m_system_filter_matched{ true };
+  bool m_after_empty_context_block{ };
 };
