@@ -2,6 +2,8 @@
 
 #include <string>
 #include <cctype>
+#include <optional>
+#include <stdexcept>
 
 template<typename ForwardIt>
 bool skip(ForwardIt* it, ForwardIt end, const char* str) {
@@ -17,6 +19,12 @@ bool skip(ForwardIt* it, ForwardIt end, const char* str) {
 }
 
 template<typename ForwardIt>
+bool skip(ForwardIt* it, ForwardIt end, char c) {
+  const char mark[2] = { c, '\0' };
+  return skip(it, end, mark);
+}
+
+template<typename ForwardIt>
 bool skip_until(ForwardIt* it, ForwardIt end, const char* str) {
   while (*it != end) {
     if (skip(it, end, str))
@@ -24,6 +32,32 @@ bool skip_until(ForwardIt* it, ForwardIt end, const char* str) {
     ++(*it);
   }
   return false;
+}
+
+template<typename ForwardIt>
+bool skip_until(ForwardIt* it, ForwardIt end, char c) {
+  const char mark[2] = { c, '\0' };
+  return skip_until(it, end, mark);
+}
+
+template<typename ForwardIt>
+bool skip_until_not_in_string(ForwardIt* it, ForwardIt end, const char* str) {
+  while (*it != end) {
+    if (skip(it, end, '"') || skip(it, end, '\'')) {
+      if (!skip_until(it, end, *std::prev(*it)))
+        throw std::runtime_error("Unterminated string");
+    }
+    if (skip(it, end, str))
+      return true;
+    ++(*it);
+  }
+  return false;
+}
+
+template<typename ForwardIt>
+bool skip_until_not_in_string(ForwardIt* it, ForwardIt end, char c) {
+  const char mark[2] = { c, '\0' };
+  return skip_until_not_in_string(it, end, mark);
 }
 
 template<typename ForwardIt>
@@ -73,6 +107,13 @@ bool trim_comment(ForwardIt it, ForwardIt* end) {
 }
 
 template<typename ForwardIt>
+void trim_space(ForwardIt begin, ForwardIt* end) {
+  while (*end != begin &&
+         std::isspace(static_cast<unsigned char>(*std::prev(*end))))
+    --(*end);
+}
+
+template<typename ForwardIt>
 void skip_value(ForwardIt* it, ForwardIt end) {
   while (*it != end && 
      (std::isalnum(static_cast<unsigned char>(**it)) || 
@@ -93,10 +134,9 @@ template<typename ForwardIt>
 std::string read_value(ForwardIt* it, ForwardIt end) {
   const auto begin = *it;
   if (skip(it, end, "'") || skip(it, end, "\"")) {
-    const char mark[2] = { *(*it - 1), '\0' };
-    if (skip_until(it, end, mark))
-      return std::string(begin + 1, *it - 1);
-    return std::string();
+    if (!skip_until(it, end, *std::prev(*it)))
+      throw std::runtime_error("Unterminated string");
+    return std::string(std::next(begin), std::prev(*it));
   }
   skip_value(it, end);
   return std::string(begin, *it);
@@ -110,9 +150,21 @@ std::string read_ident(ForwardIt* it, ForwardIt end) {
 }
 
 template<typename ForwardIt>
-int read_number(ForwardIt* it, ForwardIt end) {
+std::optional<int> try_read_number(ForwardIt* it, ForwardIt end) {
+  const auto begin = *it;
   auto number = 0;
   for (; *it != end && **it >= '0' && **it <= '9'; ++*it)
     number = number * 10 + (**it - '0');
-  return number;
+  return (*it != begin ? std::make_optional(number) : std::nullopt);
+}
+
+template<typename ForwardIt>
+bool skip_ident_with_arglist(ForwardIt* it, ForwardIt end) {
+  const auto begin = *it;
+  skip_ident(it, end);
+  if (*it != begin)
+    if (skip(it, end, "["))
+      if (!skip_until(it, end, "]"))
+        return false;
+  return true;
 }
