@@ -1,12 +1,14 @@
 #pragma once
 
 #include <string>
+#include <cstdint>
 
 // carefully selected to match platform's scancodes
-// https://developer.mozilla.org/en-US/docs/Web/API/KeyboardEvent/code
+// https://developer.mozilla.org/en-US/docs/Web/API/UI_Events/Keyboard_event_code_values
 enum class Key : uint16_t {
   none               = 0,
 
+#if defined(__linux__) || defined(_WIN32)
   Escape             = 0x0001,
   Digit1             = 0x0002,
   Digit2             = 0x0003,
@@ -92,6 +94,7 @@ enum class Key : uint16_t {
   IntlBackslash      = 0x0056,
   F11                = 0x0057,
   F12                = 0x0058,
+#endif // defined(__linux__) || defined(_WIN32)
 
 #if defined(__linux__)
   NumLock            = 0x0045,
@@ -130,27 +133,31 @@ enum class Key : uint16_t {
   MetaRight          = 0x007E,
   ContextMenu        = 0x007F,
   BrowserStop        = 0x0080,
-//Again              = 0x0081,
-//Props              = 0x0082,
-//Undo               = 0x0083,
-//Select             = 0x0084,
-//Copy               = 0x0085,
-//Open               = 0x0086,
-//Paste              = 0x0087,
-//Find               = 0x0088,
-//Cut                = 0x0089,
-//Help               = 0x008A,
-//Sleep              = 0x008E,
-//WakeUp             = 0x008F,
+  Again              = 0x0081,
+  Props              = 0x0082,
+  Undo               = 0x0083,
+  Select             = 0x0084,
+  Copy               = 0x0085,
+  Open               = 0x0086,
+  Paste              = 0x0087,
+  Find               = 0x0088,
+  Cut                = 0x0089,
+  Help               = 0x008A,
+  LaunchApp2         = 0x008C,
+  Sleep              = 0x008E,
+  WakeUp             = 0x008F,
   LaunchApp1         = 0x0090,
+  LaunchMail         = 0x009B,
   BrowserFavorites   = 0x009C,
   BrowserBack        = 0x009E,
   BrowserForward     = 0x009F,
-//Eject              = 0x00A1,
+  Eject              = 0x00A1,
   MediaTrackNext     = 0x00A3,
   MediaPlayPause     = 0x00A4,
   MediaTrackPrevious = 0x00A5,
   MediaStop          = 0x00A6,
+  LaunchMediaPlayer  = 0x00AB,
+  BrowserHome        = 0x00AC,
   BrowserRefresh     = 0x00AD,
   F13                = 0x00B7,
   F14                = 0x00B8,
@@ -170,7 +177,7 @@ enum class Key : uint16_t {
   BrightnessUp       = 0x00E1,
   DisplayToggleIntExt= 0x00E3,
   WLAN               = 0x00EE,
-//Fn                 = 0x01D0,
+  Fn                 = 0x01D0,
 #endif // defined(__linux__)
 
 #if defined(_WIN32)
@@ -206,12 +213,12 @@ enum class Key : uint16_t {
   NumpadEnter        = 0xE01C,
   ControlRight       = 0xE01D,
   AudioVolumeMute    = 0xE020,
-//LaunchApp2         = 0xE021,
+  LaunchApp2         = 0xE021,
   MediaPlayPause     = 0xE022,
   MediaStop          = 0xE024,
   AudioVolumeDown    = 0xE02E,
   AudioVolumeUp      = 0xE030,
-//BrowserHome        = 0xE032,
+  BrowserHome        = 0xE032,
   Prog3              = 0xE033,
   NumpadDivide       = 0xE035,
   PrintScreen        = 0xE037,
@@ -240,9 +247,197 @@ enum class Key : uint16_t {
   BrowserForward     = 0xE069,
   BrowserBack        = 0xE06A,
   LaunchApp1         = 0xE06B,
-//LaunchMail         = 0xE06C,
-//LaunchMediaPlayer  = 0xE06D,
+  LaunchMail         = 0xE06C,
+  LaunchMediaPlayer  = 0xE06D,
+
+  // TODO:
+  Again              = 0xD001,
+  Props              = 0xD002,
+  Undo               = 0xD003,
+  Select             = 0xD004,
+  Copy               = 0xD005,
+  Open               = 0xD006,
+  Paste              = 0xD007,
+  Find               = 0xD008,
+  Cut                = 0xD009,
+  Help               = 0xD010,
+  Sleep              = 0xD011,
+  WakeUp             = 0xD012,
+  Eject              = 0xD013,
+  Fn                 = 0xD014,
 #endif // defined(_WIN32)
+
+#if defined(__APPLE__)
+  KeyA               = 0x04, A = KeyA,
+  KeyB               = 0x05, B = KeyB,
+  KeyC               = 0x06, C = KeyC,
+  KeyD               = 0x07, D = KeyD,
+  KeyE               = 0x08, E = KeyE,
+  KeyF               = 0x09, F = KeyF,
+  KeyG               = 0x0A, G = KeyG,
+  KeyH               = 0x0B, H = KeyH,
+  KeyI               = 0x0C, I = KeyI,
+  KeyJ               = 0x0D, J = KeyJ,
+  KeyK               = 0x0E, K = KeyK,
+  KeyL               = 0x0F, L = KeyL,
+  KeyM               = 0x10, M = KeyM,
+  KeyN               = 0x11, N = KeyN,
+  KeyO               = 0x12, O = KeyO,
+  KeyP               = 0x13, P = KeyP,
+  KeyQ               = 0x14, Q = KeyQ,
+  KeyR               = 0x15, R = KeyR,
+  KeyS               = 0x16, S = KeyS,
+  KeyT               = 0x17, T = KeyT,
+  KeyU               = 0x18, U = KeyU,
+  KeyV               = 0x19, V = KeyV,
+  KeyW               = 0x1A, W = KeyW,
+  KeyX               = 0x1B, X = KeyX,
+  KeyY               = 0x1C, Y = KeyY,
+  KeyZ               = 0x1D, Z = KeyZ,
+  Digit1             = 0x1E,
+  Digit2             = 0x1F,
+  Digit3             = 0x20,
+  Digit4             = 0x21,
+  Digit5             = 0x22,
+  Digit6             = 0x23,
+  Digit7             = 0x24,
+  Digit8             = 0x25,
+  Digit9             = 0x26,
+  Digit0             = 0x27,
+  Enter              = 0x28,
+  Escape             = 0x29,
+  Backspace          = 0x2A,
+  Tab                = 0x2B,
+  Space              = 0x2C,
+  Minus              = 0x2D,
+  Equal              = 0x2E,
+  BracketLeft        = 0x2F,
+  BracketRight       = 0x30,
+  Backslash          = 0x31,
+  IntlRo             = 0x32,
+  Semicolon          = 0x33,
+  Quote              = 0x34,
+  Backquote          = 0x35,
+  Comma              = 0x36,
+  Period             = 0x37,
+  Slash              = 0x38,
+  CapsLock           = 0x39,
+  F1                 = 0x3A,
+  F2                 = 0x3B,
+  F3                 = 0x3C,
+  F4                 = 0x3D,
+  F5                 = 0x3E,
+  F6                 = 0x3F,
+  F7                 = 0x40,
+  F8                 = 0x41,
+  F9                 = 0x42,
+  F10                = 0x43,
+  F11                = 0x44,
+  F12                = 0x45,
+  PrintScreen        = 0x46,
+  ScrollLock         = 0x47,
+  Pause              = 0x48,
+  Insert             = 0x49,
+  Home               = 0x4A,
+  PageUp             = 0x4B,
+  Delete             = 0x4C,
+  End                = 0x4D,
+  PageDown           = 0x4E,
+  ArrowRight         = 0x4F,
+  ArrowLeft          = 0x50,
+  ArrowDown          = 0x51,
+  ArrowUp            = 0x52,
+  NumLock            = 0x53,
+  NumpadDivide       = 0x54,
+  NumpadMultiply     = 0x55,
+  NumpadSubtract     = 0x56,
+  NumpadAdd          = 0x57,
+  NumpadEnter        = 0x58,
+  Numpad1            = 0x59,
+  Numpad2            = 0x5A,
+  Numpad3            = 0x5B,
+  Numpad4            = 0x5C,
+  Numpad5            = 0x5D,
+  Numpad6            = 0x5E,
+  Numpad7            = 0x5F,
+  Numpad8            = 0x60,
+  Numpad9            = 0x61,
+  Numpad0            = 0x62,
+  NumpadDecimal      = 0x63,
+  IntlBackslash      = 0x64,
+  LaunchApp1         = 0x65,
+  Power              = 0x66,
+  NumpadEqual        = 0x67,
+  F13                = 0x68,
+  F14                = 0x69,
+  F15                = 0x6A,
+  F16                = 0x6B,
+  F17                = 0x6C,
+  F18                = 0x6D,
+  F19                = 0x6E,
+  F20                = 0x6F,
+  F21                = 0x70,
+  F22                = 0x71,
+  F23                = 0x72,
+  F24                = 0x73,
+  AudioVolumeMute    = 0x7F,
+  AudioVolumeUp      = 0x80,
+  AudioVolumeDown    = 0x81,
+  NumpadComma        = 0x85,
+  Lang1              = 0x90,
+  Lang2              = 0x91,
+  ControlLeft        = 0xE0,
+  ShiftLeft          = 0xE1,
+  AltLeft            = 0xE2,
+  MetaLeft           = 0xE3,
+  ControlRight       = 0xE4,
+  ShiftRight         = 0xE5,
+  AltRight           = 0xE6,
+  MetaRight          = 0xE7,
+
+  MediaPlayPause     = 0xB0,
+  MediaTrackNext     = 0xB5,
+  MediaTrackPrevious = 0xB6,
+  MediaStop          = 0xB7,
+
+  // TODO:
+  BrightnessDown     = 0xD000,
+  BrightnessUp       = 0xD001,
+  BrowserBack        = 0xD002,
+  BrowserFavorites   = 0xD003,
+  BrowserForward     = 0xD004,
+  BrowserRefresh     = 0xD005,
+  BrowserSearch      = 0xD006,
+  BrowserStop        = 0xD007,
+  Cancel             = 0xD008,
+  ContextMenu        = 0xD009,
+  Convert            = 0xD010,
+  DisplayToggleIntExt= 0xD011,
+  IntlYen            = 0xD012,
+  KanaMode           = 0xD013,
+  NonConvert         = 0xD014,
+  Prog3              = 0xD015,
+  Settings           = 0xD016,
+  WLAN               = 0xD017,
+  LaunchApp2         = 0xD018,
+  LaunchMail         = 0xD019,
+  LaunchMediaPlayer  = 0xD020,
+  BrowserHome        = 0xD021,
+  Again              = 0xD022,
+  Props              = 0xD023,
+  Undo               = 0xD024,
+  Select             = 0xD025,
+  Copy               = 0xD026,
+  Open               = 0xD027,
+  Paste              = 0xD028,
+  Find               = 0xD029,
+  Cut                = 0xD030,
+  Help               = 0xD031,
+  Sleep              = 0xD032,
+  WakeUp             = 0xD033,
+  Eject              = 0xD034,
+  Fn                 = 0xD035,
+#endif // defined(__APPLE__)
 
   ButtonLeft         = 0x0110,
   ButtonRight        = 0x0111,
@@ -252,27 +447,38 @@ enum class Key : uint16_t {
 
   any                = 0xF000,
   timeout            = 0xF001,
+  ContextActive      = 0xF002,
+
+  last_keyboard_key  = 0xEFFF,
   first_virtual      = 0xF100,
+  last_virtual       = 0xF1FF,
   first_action       = 0xF200,
+  last_action        = 0xF2FF,
   first_logical      = 0xF300,
-  last_logical       = 0xF400,
+  last_logical       = 0xF3FF,
 
   first_mouse_button = ButtonLeft,
   last_mouse_button  = ButtonForward,
-  last_virtual       = first_action - 1,
-  last_action        = first_logical - 1,
+
+  Shift              = first_logical,
+  Control            = first_logical + 1,
+  Meta               = first_logical + 2,
 };
 
-inline uint16_t operator*(Key key) { return static_cast<uint16_t>(key); }
+constexpr uint16_t operator*(Key key) { return static_cast<uint16_t>(key); }
 
-inline bool is_virtual_key(Key key) {
+constexpr bool is_virtual_key(Key key) {
   return (key >= Key::first_virtual && key <= Key::last_virtual);
 }
 
-inline bool is_mouse_button(Key key) {
+constexpr bool is_mouse_button(Key key) {
   return (key >= Key::first_mouse_button && key <= Key::last_mouse_button);
 }
 
-inline bool is_action_key(Key key) {
+constexpr bool is_keyboard_key(Key key) {
+  return (key > Key::none && key <= Key::last_keyboard_key && !is_mouse_button(key));
+}
+
+constexpr bool is_action_key(Key key) {
   return (key >= Key::first_action && key <= Key::last_action);
 }
