@@ -26,7 +26,7 @@ namespace {
   // not commutative, first parameter needs to be input sequence
   bool timeout_unifiable(const KeyEvent& se, const KeyEvent& ee) {
     const auto time_reached = (se.timeout >= ee.timeout);
-    const auto is_not = (se.state == KeyState::Not || ee.state == KeyState::Not);
+    const auto is_not = (is_not_timeout(se.state) || is_not_timeout(ee.state));
     return (is_not ? !time_reached : time_reached);
   }
 
