@@ -1,6 +1,5 @@
 
 #include "ClientPort.h"
-#include "runtime/Stage.h"
 
 namespace {
   KeySequence read_key_sequence(Deserializer& d) {
@@ -8,8 +7,7 @@ namespace {
     auto size = d.read<uint32_t>();
     for (auto i = 0u; i < size; ++i) {
       auto& event = sequence.emplace_back();
-      d.read(&event.key);
-      d.read(&event.data);
+      d.read(&event);
     }
     return sequence;
   }
@@ -46,6 +44,14 @@ namespace {
       // device filter
       context.device_filter.resize(d.read<uint32_t>(), ' ');
       d.read(context.device_filter.data(), context.device_filter.size());
+      d.read(&context.invert_device_filter);
+      
+      // modifier filter
+      context.modifier_filter = read_key_sequence(d);
+      d.read(&context.invert_modifier_filter);
+
+      // fallthrough
+      d.read(&context.fallthrough);
     }
     return std::make_unique<Stage>(std::move(contexts));
   }
@@ -58,33 +64,21 @@ namespace {
   }
 } // namespace
 
-ClientPort::ClientPort() = default;
-ClientPort::~ClientPort() = default;
-
-Connection::Socket ClientPort::socket() const {
-  return (m_connection ? m_connection->socket() : Connection::invalid_socket);
+ClientPort::ClientPort() 
+  : m_host("keymapper") {
 }
 
-Connection::Socket ClientPort::listen_socket() const {
-  return (m_connection ? m_connection->listen_socket() : Connection::invalid_socket);
-}
-
-bool ClientPort::initialize() {
-  auto connection = std::make_unique<Connection>();
-  if (!connection->listen())
-    return false;
-
-  m_connection = std::move(connection);
-  return true;
+bool ClientPort::listen() {
+  return m_host.listen();
 }
 
 bool ClientPort::accept() {
-  return (m_connection && m_connection->accept());
+  m_connection = m_host.accept();
+  return static_cast<bool>(m_connection);
 }
 
 void ClientPort::disconnect() {
-  if (m_connection)
-    m_connection->disconnect();
+  m_connection.disconnect();
 }
 
 std::unique_ptr<Stage> ClientPort::read_config(Deserializer& d) {
@@ -97,8 +91,60 @@ const std::vector<int>& ClientPort::read_active_contexts(Deserializer& d) {
 }
 
 bool ClientPort::send_triggered_action(int action) {
-  return m_connection && m_connection->send_message(
+  return m_connection.send_message(
     [&](Serializer& s) {
+      s.write(MessageType::execute_action);
       s.write(static_cast<uint32_t>(action));
+    });
+}
+
+bool ClientPort::send_virtual_key_state(Key key, KeyState state) {
+  return m_connection.send_message(
+    [&](Serializer& s) {
+      s.write(MessageType::virtual_key_state);
+      s.write(key);
+      s.write(state);
+    });
+}
+
+bool ClientPort::send_device_names(const std::vector<std::string>& device_names) {
+  return m_connection.send_message(
+    [&](Serializer& s) {
+      s.write(MessageType::device_names);
+      s.write(device_names.size());
+      for (const auto& device_name : device_names)
+        s.write(device_name);
+    });
+}
+
+bool ClientPort::read_messages(MessageHandler& handler,
+    std::optional<Duration> timeout) {
+  return m_connection.read_messages(timeout,
+    [&](Deserializer& d) {
+      switch (d.read<MessageType>()) {
+        case MessageType::configuration: {
+          handler.on_configuration_message(read_config(d));
+          break;
+        }
+        case MessageType::active_contexts: {
+          handler.on_active_contexts_message(read_active_contexts(d));
+          break;
+        }
+        case MessageType::set_virtual_key_state: {
+          const auto key = d.read<Key>();
+          const auto state = d.read<KeyState>();
+          handler.on_set_virtual_key_state_message(key, state);
+          break;
+        }
+        case MessageType::validate_state: {
+          handler.on_validate_state_message();
+          break;
+        }
+        case MessageType::device_names: {
+          handler.on_device_names_message();
+          break;
+        }
+        default: break;
+      }
     });
 }
