@@ -2,15 +2,20 @@
 #include "test.h"
 
 namespace {
-  template<size_t N>
-  std::string apply_input(Stage& stage, const char(&input)[N], 
+  std::string apply_input(Stage& stage, const KeySequence& input, 
                           int device_index = 0) {
     // apply_input all input events and concatenate output
     auto sequence = KeySequence();
-    for (auto event : parse_sequence(input))
+    for (auto event : input)
       for (auto output : stage.update(event, device_index))
         sequence.push_back(output);
     return format_sequence(sequence);
+  }
+
+  template<size_t N>
+  std::string apply_input(Stage& stage, const char(&input)[N], 
+                          int device_index = 0) {
+    return apply_input(stage, parse_sequence(input), device_index);
   }
 
   std::string apply_input(Stage& stage, KeyEvent event,
@@ -350,6 +355,96 @@ TEST_CASE("Any matches any key", "[Stage]") {
 
 //--------------------------------------------------------------------
 
+TEST_CASE("Not in input", "[Stage]") {
+  auto config = R"(
+    !C A >> X
+    B !B >> Y
+    D !D E >> Z
+  )";
+  Stage stage = create_stage(config);
+
+  REQUIRE(apply_input(stage, "+A") == "+X");
+  REQUIRE(apply_input(stage, "+A") == "+X");
+  REQUIRE(apply_input(stage, "-A") == "-X");
+  REQUIRE(stage.is_clear());
+
+  REQUIRE(apply_input(stage, "+C") == "+C");
+  REQUIRE(apply_input(stage, "+A") == "+A");
+  REQUIRE(apply_input(stage, "-A") == "-A");
+  REQUIRE(apply_input(stage, "-C") == "-C");
+  REQUIRE(stage.is_clear());
+
+  // output on release
+  REQUIRE(apply_input(stage, "+B") == "");
+  REQUIRE(apply_input(stage, "+B") == "");
+  REQUIRE(apply_input(stage, "-B") == "+Y -Y");
+  REQUIRE(stage.is_clear());
+
+  REQUIRE(apply_input(stage, "+D") == "");
+  REQUIRE(apply_input(stage, "-D") == "");
+  REQUIRE(apply_input(stage, "+E") == "+Z");
+  REQUIRE(apply_input(stage, "-E") == "-Z");
+  REQUIRE(stage.is_clear());
+
+  REQUIRE(apply_input(stage, "+D") == "");
+  REQUIRE(apply_input(stage, "+E") == "+D +E");
+  REQUIRE(apply_input(stage, "-E") == "-E");
+  REQUIRE(apply_input(stage, "-D") == "-D");
+  REQUIRE(stage.is_clear());
+
+  REQUIRE(apply_input(stage, "+D") == "");
+  REQUIRE(apply_input(stage, "+B") == "+D");
+  REQUIRE(apply_input(stage, "+B") == "");
+  REQUIRE(apply_input(stage, "-B") == "+Y -Y");
+  REQUIRE(apply_input(stage, "-D") == "-D");
+  REQUIRE(stage.is_clear());
+}
+
+//--------------------------------------------------------------------
+
+TEST_CASE("Not in input with modifier group", "[Stage]") {
+  auto config = R"(
+    A{B !B} >> X
+    C{D} !C >> Y
+  )";
+  Stage stage = create_stage(config);
+
+  REQUIRE(apply_input(stage, "+A") == "");
+  REQUIRE(apply_input(stage, "+B") == "");
+  REQUIRE(apply_input(stage, "-B") == "+X -X");
+  REQUIRE(apply_input(stage, "-A") == "");
+  REQUIRE(stage.is_clear());
+
+  REQUIRE(apply_input(stage, "+A") == "");
+  REQUIRE(apply_input(stage, "+B") == "");
+  // this is how it currently is, +A +B -A may be better...
+  REQUIRE(apply_input(stage, "-A") == "+A -A +B");
+  REQUIRE(apply_input(stage, "-B") == "-B");
+  REQUIRE(stage.is_clear());
+
+  REQUIRE(apply_input(stage, "+C") == "");
+  REQUIRE(apply_input(stage, "+D") == "");
+  REQUIRE(apply_input(stage, "-C") == "+Y -Y");
+  REQUIRE(apply_input(stage, "-D") == "");
+  REQUIRE(stage.is_clear());
+
+  REQUIRE(apply_input(stage, "+C") == "");
+  REQUIRE(apply_input(stage, "+D") == "");
+  REQUIRE(apply_input(stage, "-D") == "");
+  REQUIRE(apply_input(stage, "-C") == "+Y -Y");
+  REQUIRE(stage.is_clear());
+
+  REQUIRE(apply_input(stage, "+C") == "");
+  REQUIRE(apply_input(stage, "+D") == "");
+  REQUIRE(apply_input(stage, "+E") == "+C +D +E");
+  REQUIRE(apply_input(stage, "-C") == "-C");
+  REQUIRE(apply_input(stage, "-E") == "-E");
+  REQUIRE(apply_input(stage, "-D") == "-D");
+  REQUIRE(stage.is_clear());
+}
+
+//--------------------------------------------------------------------
+
 TEST_CASE("Not in output", "[Stage]") {
   auto config = R"(
     Shift    >> Shift
@@ -425,7 +520,10 @@ TEST_CASE("Toggle Virtual", "[Stage]") {
   REQUIRE(apply_input(stage, "+B") == "+B");
   REQUIRE(apply_input(stage, "-B") == "-B");
 
-  REQUIRE(apply_input(stage, "+ScrollLock") == "+X -X");
+  REQUIRE(apply_input(stage, "+ScrollLock") == "+Virtual1 -Virtual1 +X -X +Virtual2 -Virtual2");
+  // virtual keys are injected by server as a response to output
+  REQUIRE(apply_input(stage, "+Virtual1") == "");
+  REQUIRE(apply_input(stage, "+Virtual2") == "");
   REQUIRE(apply_input(stage, "-ScrollLock") == "");
   REQUIRE(format_sequence(stage.sequence()) == "#Virtual1 #Virtual2");
 
@@ -435,7 +533,10 @@ TEST_CASE("Toggle Virtual", "[Stage]") {
   REQUIRE(apply_input(stage, "+B") == "+2");
   REQUIRE(apply_input(stage, "-B") == "-2");
 
-  REQUIRE(apply_input(stage, "+ScrollLock") == "+X -X");
+  REQUIRE(apply_input(stage, "+ScrollLock") == "+Virtual1 -Virtual1 +X -X +Virtual2 -Virtual2");
+  // injected
+  REQUIRE(apply_input(stage, "-Virtual1") == "");
+  REQUIRE(apply_input(stage, "-Virtual2") == "");
   REQUIRE(apply_input(stage, "-ScrollLock") == "");
   REQUIRE(stage.is_clear());
 
@@ -483,6 +584,7 @@ TEST_CASE("Press already pressed, with Not", "[Stage]") {
   REQUIRE(apply_input(stage, "-ShiftLeft") == "-ShiftLeft");
   REQUIRE(stage.is_clear());
 }
+
 //--------------------------------------------------------------------
 
 TEST_CASE("Press already pressed, with Not 2", "[Stage]") {
@@ -758,12 +860,157 @@ TEST_CASE("Any key might match", "[Stage]") {
 
 TEST_CASE("Output on release", "[Stage]") {
   auto config = R"(
+    A  >>  X ^ Y
     MetaLeft{C} >> MetaLeft{R} ^ C M
   )";
   Stage stage = create_stage(config);
 
+  REQUIRE(apply_input(stage, "+A") == "+X -X");
+  REQUIRE(apply_input(stage, "+A") == "");
+  REQUIRE(apply_input(stage, "-A") == "+Y -Y");
+  REQUIRE(stage.is_clear());
+
   REQUIRE(apply_input(stage, "+MetaLeft") == "");
-  REQUIRE(apply_input(stage, "+C") == "+MetaLeft +R -R -MetaLeft ^ +C -C +M -M");
+  REQUIRE(apply_input(stage, "+C") == "+MetaLeft +R -R -MetaLeft");
+  REQUIRE(apply_input(stage, "+C") == "");
+  REQUIRE(apply_input(stage, "+C") == "");
+  REQUIRE(apply_input(stage, "-C") == "+C -C +M -M");
+  REQUIRE(apply_input(stage, "-MetaLeft") == "");
+  REQUIRE(stage.is_clear());
+}
+
+//--------------------------------------------------------------------
+
+TEST_CASE("Output on release at once", "[Stage]") {
+  auto config = R"(
+    A{B{C}} >> Z
+    A{B}    >> X ^ Y
+  )";
+  Stage stage = create_stage(config);
+
+  REQUIRE(apply_input(stage, "+A") == "");
+  REQUIRE(apply_input(stage, "+B") == "");
+  REQUIRE(apply_input(stage, "+C") == "+Z");
+  REQUIRE(apply_input(stage, "-A") == "");
+  REQUIRE(apply_input(stage, "-B") == "");
+  REQUIRE(apply_input(stage, "-C") == "-Z");
+  REQUIRE(stage.is_clear());
+
+  REQUIRE(apply_input(stage, "+A") == "");
+  REQUIRE(apply_input(stage, "+B") == "");
+  REQUIRE(apply_input(stage, "-B") == "+X -X +Y -Y");
+  REQUIRE(apply_input(stage, "-A") == "");
+  REQUIRE(stage.is_clear());
+}
+
+//--------------------------------------------------------------------
+
+TEST_CASE("Interleaving output on release", "[Stage]") {
+  auto config = R"(
+    ButtonForward >> U ^ M
+    ButtonBack    >> N ^ P
+  )";
+  Stage stage = create_stage(config);
+
+  REQUIRE(apply_input(stage, "+ButtonForward") == "+U -U");
+  REQUIRE(apply_input(stage, "+ButtonForward") == "");
+  REQUIRE(apply_input(stage, "-ButtonForward") == "+M -M");
+  REQUIRE(stage.is_clear());
+
+  REQUIRE(apply_input(stage, "+ButtonForward") == "+U -U");
+  REQUIRE(apply_input(stage, "+ButtonForward") == "");
+  REQUIRE(apply_input(stage, "+A") == "+A");
+  REQUIRE(apply_input(stage, "-A") == "-A");
+  REQUIRE(apply_input(stage, "-ButtonForward") == "+M -M");
+  REQUIRE(stage.is_clear());
+
+  REQUIRE(apply_input(stage, "+ButtonForward") == "+U -U");
+  REQUIRE(apply_input(stage, "+ButtonForward") == "");
+  REQUIRE(apply_input(stage, "+ButtonBack") == "+N -N");
+  REQUIRE(apply_input(stage, "+ButtonBack") == "");
+  REQUIRE(apply_input(stage, "-ButtonForward") == "+M -M");
+  REQUIRE(apply_input(stage, "-ButtonBack") == "+P -P");
+  REQUIRE(stage.is_clear());
+}
+
+//--------------------------------------------------------------------
+
+TEST_CASE("Output on release toggle virtual", "[Stage]") {
+  auto config = R"(
+    MetaLeft       >> Virtual1 ^ Virtual1
+    Virtual1{A}    >> X
+    !Virtual1 A    >> Y
+  )";
+  Stage stage = create_stage(config);
+
+  REQUIRE(apply_input(stage, "+A") == "+Y");
+  REQUIRE(apply_input(stage, "+A") == "+Y");
+  REQUIRE(apply_input(stage, "-A") == "-Y");
+  REQUIRE(apply_input(stage, "+MetaLeft") == "+Virtual1 -Virtual1");
+  // injected
+  REQUIRE(apply_input(stage, "+Virtual1") == "");
+  REQUIRE(apply_input(stage, "+MetaLeft") == "");
+  REQUIRE(format_sequence(stage.sequence()) == "#MetaLeft #Virtual1");
+  REQUIRE(apply_input(stage, "+A") == "+X");
+  REQUIRE(apply_input(stage, "+A") == "+X");
+  REQUIRE(apply_input(stage, "-A") == "-X");
+  REQUIRE(apply_input(stage, "-MetaLeft") == "+Virtual1 -Virtual1");
+  // injected
+  REQUIRE(apply_input(stage, "-Virtual1") == "");
+  REQUIRE(format_sequence(stage.sequence()) == "");
+  REQUIRE(apply_input(stage, "+A") == "+Y");
+  REQUIRE(apply_input(stage, "+A") == "+Y");
+  REQUIRE(apply_input(stage, "-A") == "-Y");
+  REQUIRE(stage.is_clear());
+}
+
+//--------------------------------------------------------------------
+
+TEST_CASE("Output on release with timeout", "[Stage]") {
+  auto config = R"(
+    Escape{250ms}  >>  X ^ Y
+  )";
+  Stage stage = create_stage(config);
+
+  REQUIRE(apply_input(stage, "+Escape") == "-250ms");
+  REQUIRE(apply_input(stage, reply_timeout_ms(123)) == "+Escape");
+  REQUIRE(apply_input(stage, "-Escape") == "-Escape");
+  REQUIRE(stage.is_clear());
+
+  REQUIRE(apply_input(stage, "+Escape") == "-250ms");
+  REQUIRE(apply_input(stage, reply_timeout_ms(250)) == "+X -X");
+  REQUIRE(apply_input(stage, "-Escape") == "+Y -Y");
+}
+
+//--------------------------------------------------------------------
+
+TEST_CASE("Cancelling output on release", "[Stage]") {
+  auto config = R"(
+    ButtonForward  >> $(system command 1) ^ $(system command 2)
+
+    [title = "SomeApp"]
+    ButtonMiddle >> ButtonLeft ^ M
+  )";
+  Stage stage = create_stage(config);
+
+  stage.set_active_client_contexts({ 0, 1 });
+  REQUIRE(apply_input(stage, "+ButtonMiddle") == "+ButtonLeft -ButtonLeft");
+  REQUIRE(apply_input(stage, "+ButtonMiddle") == "");
+  REQUIRE(apply_input(stage, "-ButtonMiddle") == "+M -M");
+  REQUIRE(stage.is_clear());
+
+  REQUIRE(apply_input(stage, "+ButtonMiddle") == "+ButtonLeft -ButtonLeft");
+  REQUIRE(apply_input(stage, "+ButtonMiddle") == "");
+  stage.set_active_client_contexts({ 0 }); // focus changed no longer active
+  REQUIRE(apply_input(stage, "-ButtonMiddle") == "");
+  REQUIRE(stage.is_clear());
+
+  stage.set_active_client_contexts({ 0, 1 }); // focus changed
+  REQUIRE(apply_input(stage, "+ButtonForward") == "+Action0");
+  REQUIRE(apply_input(stage, "+ButtonForward") == "");
+  stage.set_active_client_contexts({ 0 }); // focus changed but still active
+  REQUIRE(apply_input(stage, "-ButtonForward") == "+Action1 -Action1 -Action0");
+  REQUIRE(stage.is_clear());
 }
 
 //--------------------------------------------------------------------
@@ -858,7 +1105,7 @@ TEST_CASE("Mapping sequence in context", "[Stage]") {
   )";
   Stage stage = create_stage(config);
   REQUIRE(stage.contexts().size() == 4);
-  stage.set_active_contexts({ 0, 3 }); // No program
+  stage.set_active_client_contexts({ 0, 3 }); // No program
 
 #if defined(__linux__)
   REQUIRE(apply_input(stage, "+A -A") == "+E -E");
@@ -868,12 +1115,12 @@ TEST_CASE("Mapping sequence in context", "[Stage]") {
   REQUIRE(apply_input(stage, "+R -R") == "+R -R");
   REQUIRE(apply_input(stage, "+X -X") == "+X -X"); // implicit default mapping forwards
 
-  stage.set_active_contexts({ 0, 1, 3 }); // Firefox
+  stage.set_active_client_contexts({ 0, 1, 3 }); // Firefox
   REQUIRE(apply_input(stage, "+A -A") == "+B -B");
   REQUIRE(apply_input(stage, "+R -R") == "+R -R");
   REQUIRE(apply_input(stage, "+X -X") == "+Y -Y");
 
-  stage.set_active_contexts({ 0, 2, 3 }); // Konsole
+  stage.set_active_client_contexts({ 0, 2, 3 }); // Konsole
   REQUIRE(apply_input(stage, "+A -A") == "+C -C");
   REQUIRE(apply_input(stage, "+R -R") == "+R -R");
   REQUIRE(apply_input(stage, "+X -X") == "+Z -Z");
@@ -906,7 +1153,7 @@ TEST_CASE("Mapping sequence in context - comparison", "[Stage]") {
   )";
   Stage stage = create_stage(config);
   REQUIRE(stage.contexts().size() == 4);
-  stage.set_active_contexts({ 0, 1 }); // No program
+  stage.set_active_client_contexts({ 0, 1 }); // No program
 
 #if defined(__linux__)
   REQUIRE(apply_input(stage, "+A -A") == "+E -E");
@@ -916,12 +1163,12 @@ TEST_CASE("Mapping sequence in context - comparison", "[Stage]") {
   REQUIRE(apply_input(stage, "+R -R") == "+R -R");
   REQUIRE(apply_input(stage, "+X -X") == "+X -X"); // no default mapping for command3
 
-  stage.set_active_contexts({ 0, 1, 2 }); // Firefox
+  stage.set_active_client_contexts({ 0, 1, 2 }); // Firefox
   REQUIRE(apply_input(stage, "+A -A") == "+B -B");
   REQUIRE(apply_input(stage, "+R -R") == "+U -U");
   REQUIRE(apply_input(stage, "+X -X") == "+Y -Y");
 
-  stage.set_active_contexts({ 0, 1, 3 }); // Konsole
+  stage.set_active_client_contexts({ 0, 1, 3 }); // Konsole
   REQUIRE(apply_input(stage, "+A -A") == "+C -C");
   REQUIRE(apply_input(stage, "+R -R") == "+V -V");
   REQUIRE(apply_input(stage, "+X -X") == "+Z -Z");
@@ -955,7 +1202,7 @@ TEST_CASE("Restore default context", "[Stage]") {
   )";
   Stage stage = create_stage(config);
   REQUIRE(stage.contexts().size() == 5);
-  stage.set_active_contexts({ 1, 4 }); // No program
+  stage.set_active_client_contexts({ 1, 4 }); // No program
 
 #if defined(__linux__)
   REQUIRE(apply_input(stage, "+A -A") == "+E -E");
@@ -965,20 +1212,259 @@ TEST_CASE("Restore default context", "[Stage]") {
   REQUIRE(apply_input(stage, "+R -R") == "+R -R");
   REQUIRE(apply_input(stage, "+X -X") == "+X -X"); // implicit default mapping forwards
 
-  stage.set_active_contexts({ 1, 2, 4 }); // Firefox
+  stage.set_active_client_contexts({ 1, 2, 4 }); // Firefox
   REQUIRE(apply_input(stage, "+A -A") == "+B -B");
   REQUIRE(apply_input(stage, "+R -R") == "+R -R");
   REQUIRE(apply_input(stage, "+X -X") == "+Y -Y");
 
-  stage.set_active_contexts({ 1, 3, 4 }); // Konsole
+  stage.set_active_client_contexts({ 1, 3, 4 }); // Konsole
   REQUIRE(apply_input(stage, "+A -A") == "+C -C");
   REQUIRE(apply_input(stage, "+R -R") == "+R -R");
   REQUIRE(apply_input(stage, "+X -X") == "+Z -Z");
 
-  stage.set_active_contexts({ 0, 0, 1, 4 }); // AnyDesk
+  stage.set_active_client_contexts({ 0, 0, 1, 4 }); // AnyDesk
   REQUIRE(apply_input(stage, "+A -A") == "+A -A");
   REQUIRE(apply_input(stage, "+R -R") == "+R -R");
   REQUIRE(apply_input(stage, "+X -X") == "+X -X");
+}
+
+//--------------------------------------------------------------------
+
+TEST_CASE("Context with modifier filter", "[Stage]") {
+  auto config = R"(
+    [default]
+    A >> command
+    B >> Virtual1
+  
+    [modifier="!Virtual1"]
+    command >> Z
+    
+    [modifier="Virtual1 Shift"] # adds one implicit fallthrough context
+    command >> W
+  
+    [title="Firefox"]
+    command >> X
+    
+    [title="Firefox" modifier="ShiftLeft"]
+    command >> Y
+  )";
+  
+  Stage stage = create_stage(config);
+  REQUIRE(stage.contexts().size() == 6);
+  stage.set_active_client_contexts({ 0, 1, 2, 3 }); // No program
+  
+  REQUIRE(apply_input(stage, "+A -A") == "+Z -Z");
+  REQUIRE(stage.is_clear());
+  
+  REQUIRE(apply_input(stage, "+B -B") == "+Virtual1 -Virtual1");
+  // virtual keys are injected by server as a response to output
+  REQUIRE(apply_input(stage, "+Virtual1") == "");
+  REQUIRE(apply_input(stage, "+A -A") == "+A -A");
+  
+  // fallthrough contexts are added for logical keys
+  REQUIRE(apply_input(stage, "+ShiftLeft") == "+ShiftLeft");
+  REQUIRE(apply_input(stage, "+A -A") == "+W -W");
+  REQUIRE(apply_input(stage, "-ShiftLeft") == "-ShiftLeft");
+  REQUIRE(apply_input(stage, "+ShiftRight") == "+ShiftRight");
+  REQUIRE(apply_input(stage, "+A -A") == "+W -W");
+  REQUIRE(apply_input(stage, "-ShiftRight") == "-ShiftRight");
+  
+  REQUIRE(apply_input(stage, "+B -B") == "+Virtual1 -Virtual1");
+  REQUIRE(apply_input(stage, "-Virtual1") == "");
+  REQUIRE(stage.is_clear());
+
+  stage.set_active_client_contexts({ 0, 1, 2, 3, 4, 5 }); // Firefox
+
+  REQUIRE(apply_input(stage, "+A -A") == "+X -X");
+  REQUIRE(stage.is_clear());
+
+  REQUIRE(apply_input(stage, "+ShiftLeft") == "+ShiftLeft");
+  REQUIRE(apply_input(stage, "+A") == "+Y");
+  REQUIRE(apply_input(stage, "-A") == "-Y");
+  REQUIRE(apply_input(stage, "-ShiftLeft") == "-ShiftLeft");
+  REQUIRE(stage.is_clear());
+}
+
+//--------------------------------------------------------------------
+
+TEST_CASE("Context with modifier filter #2", "[Stage]") {
+  auto config = R"(
+    [modifier='A']
+    E >> F
+
+    [default]
+    A >> B
+  )";
+  
+  Stage stage = create_stage(config);
+  REQUIRE(stage.contexts().size() == 2);
+  
+  REQUIRE(apply_input(stage, "+A -A") == "+B -B");
+  REQUIRE(apply_input(stage, "+E -E") == "+E -E");
+  REQUIRE(apply_input(stage, "+A +E -E -A") == "+B +F -F -B");
+  REQUIRE(stage.is_clear());
+}
+
+//--------------------------------------------------------------------
+
+TEST_CASE("Context with modifier filter and ContextActive mapping", "[Stage]") {
+  auto config = R"(
+    [default]
+    Virtual1{ U } >> R
+    Virtual2{ U } >> S
+    U >> V
+
+    [modifier="A"]
+    ContextActive >> X
+
+    [modifier="B"]
+    ContextActive >> Virtual1
+
+    [modifier="C"]
+    U >> W
+
+    [modifier="D"]
+    ContextActive >> Virtual2 ^ Virtual2
+  )";
+  
+  Stage stage = create_stage(config, false);
+  REQUIRE(stage.contexts().size() == 5);
+  CHECK(format_sequence(stage.set_active_client_contexts({ 0, 1, 2, 3, 4 })) == "");
+  
+  CHECK(apply_input(stage, "+A") == "+X +A");
+  CHECK(apply_input(stage, "-A") == "-A -X");
+  CHECK(apply_input(stage, "+U") == "+V");
+  CHECK(apply_input(stage, "-U") == "-V");
+
+  // virtual keys are toggled on press
+  CHECK(apply_input(stage, "+B") == "+Virtual1 +B");
+  CHECK(apply_input(stage, "+Virtual1") == "");
+  CHECK(apply_input(stage, "-B") == "-B -Virtual1");
+  CHECK(apply_input(stage, "+U") == "+R");
+  CHECK(apply_input(stage, "-U") == "-R");
+  CHECK(apply_input(stage, "+B") == "+Virtual1 +B");
+  CHECK(apply_input(stage, "-Virtual1") == "");
+  CHECK(apply_input(stage, "-B") == "-B -Virtual1");
+  CHECK(apply_input(stage, "+U") == "+V");
+  CHECK(apply_input(stage, "-U") == "-V");
+
+  // toggle again on release
+  CHECK(apply_input(stage, "+D") == "+Virtual2 -Virtual2 +D");
+  CHECK(apply_input(stage, "+Virtual2") == "");
+  CHECK(apply_input(stage, "+U") == "+S");
+  CHECK(apply_input(stage, "-U") == "-S");
+  CHECK(apply_input(stage, "-D") == "-D +Virtual2 -Virtual2");
+  CHECK(apply_input(stage, "-Virtual2") == "");
+  CHECK(apply_input(stage, "+U") == "+V");
+  CHECK(apply_input(stage, "-U") == "-V");
+
+  REQUIRE(stage.is_clear());
+}
+
+//--------------------------------------------------------------------
+
+TEST_CASE("Initially active contexts and ContextActive mapping", "[Stage]") {
+  auto config = R"(
+    [default]
+    ContextActive >> Y
+
+    [modifier="!A"]
+    ContextActive >> X
+  )";
+  
+  Stage stage = create_stage(config, false);
+
+  // A is initially not hold - context is active
+  CHECK(format_sequence(stage.set_active_client_contexts({ 0, 1 })) == "+Y +X");
+
+  // +A context is toggled
+  CHECK(apply_input(stage, "+A") == "-X +A");
+
+  // -A context is toggled
+  CHECK(apply_input(stage, "-A") == "-A +X");
+}
+
+//--------------------------------------------------------------------
+
+TEST_CASE("Initially active contexts and ContextActive mapping #2", "[Stage]") {
+  auto config = R"(
+    [default]
+    ContextActive >> Y ^ Z
+
+    [modifier="!A"]
+    ContextActive >> X ^ W
+  )";
+
+  Stage stage = create_stage(config, false);
+
+  // A is initially not hold - context is active
+  CHECK(format_sequence(stage.set_active_client_contexts({ 0, 1 })) == "+Y -Y +X -X");
+
+  // +A context is toggled
+  CHECK(apply_input(stage, "+A") == "+W -W +A");
+
+  // -A context is toggled
+  CHECK(apply_input(stage, "-A") == "-A +X -X");
+}
+
+//--------------------------------------------------------------------
+
+TEST_CASE("Focusing window with ContextActive mapping", "[Stage]") {
+  auto config = R"(
+    [title="Thunar"]
+    ContextActive >> A ^ B
+
+    [title="Firefox"]
+    ContextActive >> C ^ D
+  )";
+  
+  Stage stage = create_stage(config, false);
+  REQUIRE(stage.contexts().size() == 2);
+
+  // focus first
+  CHECK(format_sequence(stage.set_active_client_contexts({ 0 })) == "+A -A");
+
+  // focus seconds
+  CHECK(format_sequence(stage.set_active_client_contexts({ 1 })) == "+B -B +C -C");
+
+  // focus first again
+  CHECK(format_sequence(stage.set_active_client_contexts({ 0 })) == "+D -D +A -A");
+
+  // focus something else
+  CHECK(format_sequence(stage.set_active_client_contexts({ })) == "+B -B");
+
+  REQUIRE(stage.is_clear());
+}
+
+//--------------------------------------------------------------------
+
+TEST_CASE("Fallthrough contexts", "[Stage]") {
+  auto config = R"(
+    [title="Thunar"]
+    [class="thunar"]
+    A >> X
+
+    [title="Firefox"]
+    [system="Linux"]   # two of
+    [system="Windows"] # three
+    [system="MacOS"]   # are removed
+    B >> Y
+  )";
+  
+  Stage stage = create_stage(config, false);
+  REQUIRE(stage.contexts().size() == 4);
+
+  CHECK(format_sequence(stage.set_active_client_contexts({ 0 })) == "");
+  CHECK(apply_input(stage, "+A +B -A -B") == "+X +B -X -B");
+  CHECK(format_sequence(stage.set_active_client_contexts({ 1 })) == "");
+  CHECK(apply_input(stage, "+A +B -A -B") == "+X +B -X -B");
+
+  CHECK(format_sequence(stage.set_active_client_contexts({ 2 })) == "");
+  CHECK(apply_input(stage, "+A +B -A -B") == "+A +Y -A -Y");
+  CHECK(format_sequence(stage.set_active_client_contexts({ 3 })) == "");
+  CHECK(apply_input(stage, "+A +B -A -B") == "+A +Y -A -Y");
+
+  REQUIRE(stage.is_clear());
 }
 
 //--------------------------------------------------------------------
@@ -999,9 +1485,11 @@ TEST_CASE("Trigger action", "[Stage]") {
   CHECK(apply_input(stage, "-B") == "-B -Action1");
   CHECK(apply_input(stage, "+C") == "+E +F -F -E +Action2 +G +H -H -G");
   CHECK(apply_input(stage, "-C") == "-Action2");
-  CHECK(apply_input(stage, "+D") == "^ +Action3");
-  CHECK(apply_input(stage, "-D") == "-Action3");
-  CHECK(apply_input(stage, "+E") == "+Action4 ^");
+  CHECK(apply_input(stage, "+D") == "");
+  CHECK(apply_input(stage, "+D") == "");
+  CHECK(apply_input(stage, "-D") == "+Action3 -Action3");
+  CHECK(apply_input(stage, "+E") == "+Action4");
+  CHECK(apply_input(stage, "+E") == "");
   CHECK(apply_input(stage, "-E") == "-Action4");
 }
 
@@ -1102,8 +1590,8 @@ TEST_CASE("New common modifier behaviour", "[Stage]") {
 
   REQUIRE(apply_input(stage, "+AltLeft") == "");
   REQUIRE(apply_input(stage, "+ControlLeft") == "+AltLeft");
-  REQUIRE(apply_input(stage, "-ControlLeft") == "-AltLeft");
-  REQUIRE(apply_input(stage, "-AltLeft") == "");
+  REQUIRE(apply_input(stage, "-ControlLeft") == "");
+  REQUIRE(apply_input(stage, "-AltLeft") == "-AltLeft");
   REQUIRE(stage.is_clear());
 
   REQUIRE(apply_input(stage, "+ControlLeft") == "+ControlLeft");
@@ -1142,29 +1630,29 @@ TEST_CASE("Continue matching after failed might-match", "[Stage]") {
   // when might-match fails, look for longest exact match
   REQUIRE(apply_input(stage, "+A") == "");
   REQUIRE(apply_input(stage, "+S") == "+X +S");
-  REQUIRE(apply_input(stage, "-S") == "-S -X");
-  REQUIRE(apply_input(stage, "-A") == "");
+  REQUIRE(apply_input(stage, "-S") == "-S");
+  REQUIRE(apply_input(stage, "-A") == "-X");
 
   REQUIRE(apply_input(stage, "+A") == "");
   REQUIRE(apply_input(stage, "+B") == "");
   REQUIRE(apply_input(stage, "+S") == "+Y +S");
-  REQUIRE(apply_input(stage, "-S") == "-S -Y");
-  REQUIRE(apply_input(stage, "-B") == "");
+  REQUIRE(apply_input(stage, "-S") == "-S");
+  REQUIRE(apply_input(stage, "-B") == "-Y");
   REQUIRE(apply_input(stage, "-A") == "");
 
   REQUIRE(apply_input(stage, "+A") == "");
   REQUIRE(apply_input(stage, "+D") == "");
   REQUIRE(apply_input(stage, "+S") == "+X +D +S");
-  REQUIRE(apply_input(stage, "-S") == "-S -X");
+  REQUIRE(apply_input(stage, "-S") == "-S");
   REQUIRE(apply_input(stage, "-D") == "-D");
-  REQUIRE(apply_input(stage, "-A") == "");
+  REQUIRE(apply_input(stage, "-A") == "-X");
 
   REQUIRE(apply_input(stage, "+A") == "");
   REQUIRE(apply_input(stage, "+D") == "");
   REQUIRE(apply_input(stage, "+F") == "+X +U");
-  REQUIRE(apply_input(stage, "-F") == "-U -X");
+  REQUIRE(apply_input(stage, "-F") == "-U");
   REQUIRE(apply_input(stage, "-D") == "");
-  REQUIRE(apply_input(stage, "-A") == "");
+  REQUIRE(apply_input(stage, "-A") == "-X");
 }
 
 //--------------------------------------------------------------------
@@ -1175,36 +1663,36 @@ TEST_CASE("Timeout Hold", "[Stage]") {
   )";
   Stage stage = create_stage(config);
 
-  CHECK(apply_input(stage, "+A") == "1000ms");
-  CHECK(apply_input(stage, make_timeout_ms(271)) == "+A");
+  CHECK(apply_input(stage, "+A") == "-1000ms");
+  CHECK(apply_input(stage, reply_timeout_ms(271)) == "+A");
   CHECK(apply_input(stage, "+B") == "+B");
   CHECK(apply_input(stage, "-B") == "-B");
   CHECK(apply_input(stage, "-A") == "-A");
   REQUIRE(stage.is_clear());
 
-  CHECK(apply_input(stage, "+A") == "1000ms");
-  CHECK(apply_input(stage, make_timeout_ms(1000)) == "+X");
-  CHECK(apply_input(stage, "+A") == "1000ms");
-  CHECK(apply_input(stage, make_timeout_ms(1000)) == "+X");
-  CHECK(apply_input(stage, "+A") == "1000ms");
+  CHECK(apply_input(stage, "+A") == "-1000ms");
+  CHECK(apply_input(stage, reply_timeout_ms(1000)) == "+X");
+  CHECK(apply_input(stage, "+A") == "-1000ms");
+  CHECK(apply_input(stage, reply_timeout_ms(1000)) == "+X");
+  CHECK(apply_input(stage, "+A") == "-1000ms");
   // release while waiting for timeout - send elapsed time
   // output is suppressed when timeout matched once
-  CHECK(apply_input(stage, make_timeout_ms(271)) == "");
+  CHECK(apply_input(stage, reply_timeout_ms(271)) == "");
   CHECK(apply_input(stage, "-A") == "-X");
   REQUIRE(stage.is_clear());
 
   CHECK(apply_input(stage, "+ShiftLeft") == "+ShiftLeft");
-  CHECK(apply_input(stage, "+A") == "1000ms");
+  CHECK(apply_input(stage, "+A") == "-1000ms");
   // release while waiting for timeout - send elapsed time
-  CHECK(apply_input(stage, make_timeout_ms(271)) == "+A");
+  CHECK(apply_input(stage, reply_timeout_ms(271)) == "+A");
   CHECK(apply_input(stage, "-A") == "-A");
   CHECK(apply_input(stage, "-ShiftLeft") == "-ShiftLeft");
   REQUIRE(stage.is_clear());
 
   CHECK(apply_input(stage, "+ShiftLeft") == "+ShiftLeft");
-  CHECK(apply_input(stage, "+A") == "1000ms");
+  CHECK(apply_input(stage, "+A") == "-1000ms");
   // another event while waiting for timeout - send elapsed time
-  CHECK(apply_input(stage, make_timeout_ms(271)) == "+A");
+  CHECK(apply_input(stage, reply_timeout_ms(271)) == "+A");
   CHECK(apply_input(stage, "-ShiftLeft") == "-ShiftLeft");
   CHECK(apply_input(stage, "-A") == "-A");
   REQUIRE(stage.is_clear());
@@ -1220,26 +1708,26 @@ TEST_CASE("Timeout Hold, Another mapping", "[Stage]") {
   Stage stage = create_stage(config);
 
   // key press which cancels timeout is currently the trigger
-  CHECK(apply_input(stage, "+A") == "1000ms");
-  CHECK(apply_input(stage, make_timeout_ms(271)) == "+Z");
+  CHECK(apply_input(stage, "+A") == "-1000ms");
+  CHECK(apply_input(stage, reply_timeout_ms(271)) == "+Z");
   CHECK(apply_input(stage, "+B") == "+B");
   CHECK(apply_input(stage, "-B") == "-B");
   CHECK(apply_input(stage, "-A") == "-Z");
   REQUIRE(stage.is_clear());
 
-  CHECK(apply_input(stage, "+A") == "1000ms");
-  CHECK(apply_input(stage, make_timeout_ms(1000)) == "+X");
-  CHECK(apply_input(stage, "+A") == "1000ms");
-  CHECK(apply_input(stage, make_timeout_ms(1000)) == "+X");
-  CHECK(apply_input(stage, "+A") == "1000ms");
+  CHECK(apply_input(stage, "+A") == "-1000ms");
+  CHECK(apply_input(stage, reply_timeout_ms(1000)) == "+X");
+  CHECK(apply_input(stage, "+A") == "-1000ms");
+  CHECK(apply_input(stage, reply_timeout_ms(1000)) == "+X");
+  CHECK(apply_input(stage, "+A") == "-1000ms");
   // release while waiting for timeout - send elapsed time
   // output is suppressed when timeout matched once
-  CHECK(apply_input(stage, make_timeout_ms(271)) == "");
+  CHECK(apply_input(stage, reply_timeout_ms(271)) == "");
   CHECK(apply_input(stage, "-A") == "-X");
   REQUIRE(stage.is_clear());
 
-  CHECK(apply_input(stage, "+A") == "1000ms");
-  CHECK(apply_input(stage, make_timeout_ms(271)) == "+Z");
+  CHECK(apply_input(stage, "+A") == "-1000ms");
+  CHECK(apply_input(stage, reply_timeout_ms(271)) == "+Z");
   CHECK(apply_input(stage, "-A") == "-Z");
   REQUIRE(stage.is_clear());
 }
@@ -1248,14 +1736,15 @@ TEST_CASE("Timeout Hold, Another mapping", "[Stage]") {
 
 TEST_CASE("Timeout", "[Stage]") {
   auto config = R"(
-    A 1000ms >> X
+    A !A 1000ms >> X
+    E 1000ms >> Y
   )";
   Stage stage = create_stage(config);
 
   CHECK(apply_input(stage, "+A") == "");
   CHECK(apply_input(stage, "+A") == "");
   CHECK(apply_input(stage, "+A") == "");
-  CHECK(apply_input(stage, "-A") == "1000ms");
+  CHECK(apply_input(stage, "-A") == "+1000ms");
   CHECK(apply_input(stage, "+B") == "+A -A +B");
   CHECK(apply_input(stage, "-B") == "-B");
   REQUIRE(stage.is_clear());
@@ -1263,8 +1752,47 @@ TEST_CASE("Timeout", "[Stage]") {
   CHECK(apply_input(stage, "+A") == "");
   CHECK(apply_input(stage, "+A") == "");
   CHECK(apply_input(stage, "+A") == "");
-  CHECK(apply_input(stage, "-A") == "1000ms");
-  CHECK(apply_input(stage, make_timeout_ms(1000)) == "+X -X");
+  CHECK(apply_input(stage, "-A") == "+1000ms");
+  CHECK(apply_input(stage, reply_timeout_ms(1000)) == "+X -X");
+  REQUIRE(stage.is_clear());
+
+  // E then interruption by F
+  CHECK(apply_input(stage, "+E") == "+1000ms");
+  CHECK(apply_input(stage, reply_timeout_ms(999)) == "+E");
+  CHECK(apply_input(stage, "+F") == "+F");
+  CHECK(apply_input(stage, "-F") == "-F");
+  CHECK(apply_input(stage, "-E") == "-E");
+  REQUIRE(stage.is_clear());
+
+  // E then interruption by E
+  CHECK(apply_input(stage, "+E") == "+1000ms");
+  CHECK(apply_input(stage, "-E") == "");
+  CHECK(apply_input(stage, reply_timeout_ms(999)) == "+E -E");
+  CHECK(apply_input(stage, "+E") == "+1000ms");
+  CHECK(apply_input(stage, "-E") == "");
+  CHECK(apply_input(stage, reply_timeout_ms(1000)) == "+Y -Y");
+  REQUIRE(stage.is_clear());
+
+  // keyrepeat is ignored while waiting for timeout
+  CHECK(apply_input(stage, "+E") == "+1000ms");
+  CHECK(apply_input(stage, reply_timeout_ms(1000)) == "+Y");
+  CHECK(apply_input(stage, "-E") == "-Y");
+  REQUIRE(stage.is_clear());
+
+  CHECK(apply_input(stage, "+ShiftLeft") == "+ShiftLeft");  
+  CHECK(apply_input(stage, "+E") == "+1000ms");  
+  CHECK(apply_input(stage, "-ShiftLeft") == "-ShiftLeft");  
+  CHECK(apply_input(stage, reply_timeout_ms(999)) == "+E");
+  CHECK(apply_input(stage, "+F") == "+F");
+  CHECK(apply_input(stage, "-F") == "-F");
+  CHECK(apply_input(stage, "-E") == "-E");
+  REQUIRE(stage.is_clear());
+
+  CHECK(apply_input(stage, "+ShiftLeft") == "+ShiftLeft");  
+  CHECK(apply_input(stage, "+E") == "+1000ms");  
+  CHECK(apply_input(stage, "-ShiftLeft") == "-ShiftLeft");  
+  CHECK(apply_input(stage, reply_timeout_ms(1000)) == "+Y");
+  CHECK(apply_input(stage, "-E") == "-Y");
   REQUIRE(stage.is_clear());
 }
 
@@ -1278,37 +1806,37 @@ TEST_CASE("Timeout Switch", "[Stage]") {
   )";
   Stage stage = create_stage(config);
 
-  CHECK(apply_input(stage, "+A") == "2000ms");
-  CHECK(apply_input(stage, make_timeout_ms(2000)) == "+X");
-  CHECK(apply_input(stage, "+A") == "2000ms");
-  CHECK(apply_input(stage, make_timeout_ms(2000)) == "+X");
-  CHECK(apply_input(stage, "+A") == "2000ms");
-  CHECK(apply_input(stage, make_timeout_ms(1999)) == "");
+  CHECK(apply_input(stage, "+A") == "-2000ms");
+  CHECK(apply_input(stage, reply_timeout_ms(2000)) == "+X");
+  CHECK(apply_input(stage, "+A") == "-2000ms");
+  CHECK(apply_input(stage, reply_timeout_ms(2000)) == "+X");
+  CHECK(apply_input(stage, "+A") == "-2000ms");
+  CHECK(apply_input(stage, reply_timeout_ms(1999)) == "");
   CHECK(apply_input(stage, "-A") == "-X");
   REQUIRE(stage.is_clear());
 
-  CHECK(apply_input(stage, "+A") == "2000ms");
-  CHECK(apply_input(stage, make_timeout_ms(1999)) == "+Y");
+  CHECK(apply_input(stage, "+A") == "-2000ms");
+  CHECK(apply_input(stage, reply_timeout_ms(1999)) == "+Y");
   CHECK(apply_input(stage, "-A") == "-Y");
   REQUIRE(stage.is_clear());
 
-  CHECK(apply_input(stage, "+A") == "2000ms");
-  CHECK(apply_input(stage, make_timeout_ms(1000)) == "+Y");
+  CHECK(apply_input(stage, "+A") == "-2000ms");
+  CHECK(apply_input(stage, reply_timeout_ms(1000)) == "+Y");
   CHECK(apply_input(stage, "-A") == "-Y");
   REQUIRE(stage.is_clear());
 
-  CHECK(apply_input(stage, "+A") == "2000ms");
-  CHECK(apply_input(stage, make_timeout_ms(999)) == "+Z");
+  CHECK(apply_input(stage, "+A") == "-2000ms");
+  CHECK(apply_input(stage, reply_timeout_ms(999)) == "+Z");
   CHECK(apply_input(stage, "-A") == "-Z");
   REQUIRE(stage.is_clear());
 
-  CHECK(apply_input(stage, "+A") == "2000ms");
-  CHECK(apply_input(stage, make_timeout_ms(500)) == "+Z");
+  CHECK(apply_input(stage, "+A") == "-2000ms");
+  CHECK(apply_input(stage, reply_timeout_ms(500)) == "+Z");
   CHECK(apply_input(stage, "-A") == "-Z");
   REQUIRE(stage.is_clear());
 
-  CHECK(apply_input(stage, "+A") == "2000ms");
-  CHECK(apply_input(stage, make_timeout_ms(499)) == "+A");
+  CHECK(apply_input(stage, "+A") == "-2000ms");
+  CHECK(apply_input(stage, reply_timeout_ms(499)) == "+A");
   CHECK(apply_input(stage, "-A") == "-A");
   REQUIRE(stage.is_clear());
 }
@@ -1322,9 +1850,9 @@ TEST_CASE("Timeout Sequence #1", "[Stage]") {
   Stage stage = create_stage(config);
 
   // key repeat is ignored
-  CHECK(apply_input(stage, "+A") == "500ms");
+  CHECK(apply_input(stage, "+A") == "-500ms");
   CHECK(apply_input(stage, "+A") == "");
-  CHECK(apply_input(stage, make_timeout_ms(500)) == "");
+  CHECK(apply_input(stage, reply_timeout_ms(500)) == "");
   CHECK(apply_input(stage, "+A") == "");
   CHECK(apply_input(stage, "-A") == "");
   CHECK(apply_input(stage, "+B") == "+X");
@@ -1333,17 +1861,17 @@ TEST_CASE("Timeout Sequence #1", "[Stage]") {
   CHECK(apply_input(stage, "-B") == "-B -X");
   REQUIRE(stage.is_clear());
 
-  CHECK(apply_input(stage, "+A") == "500ms");
+  CHECK(apply_input(stage, "+A") == "-500ms");
   CHECK(apply_input(stage, "+A") == "");
-  CHECK(apply_input(stage, make_timeout_ms(300)) == "+A");
+  CHECK(apply_input(stage, reply_timeout_ms(300)) == "+A");
   CHECK(apply_input(stage, "-A") == "-A");
   CHECK(apply_input(stage, "+C") == "+C");
   CHECK(apply_input(stage, "+C") == "+C");
   CHECK(apply_input(stage, "-C") == "-C");
   REQUIRE(stage.is_clear());
 
-  CHECK(apply_input(stage, "+A") == "500ms");
-  CHECK(apply_input(stage, make_timeout_ms(500)) == "");
+  CHECK(apply_input(stage, "+A") == "-500ms");
+  CHECK(apply_input(stage, reply_timeout_ms(500)) == "");
   CHECK(apply_input(stage, "-A") == "");
   CHECK(apply_input(stage, "+C") == "+A -A +C");
   CHECK(apply_input(stage, "+C") == "+C");
@@ -1355,15 +1883,16 @@ TEST_CASE("Timeout Sequence #1", "[Stage]") {
 
 TEST_CASE("Timeout Sequence #2", "[Stage]") {
   auto config = R"(
-    A 500ms B >> X
+    A !A 500ms B >> X
+    E 500ms F >> Y
   )";
   Stage stage = create_stage(config);
 
   // key repeat is ignored
   CHECK(apply_input(stage, "+A") == "");
   CHECK(apply_input(stage, "+A") == "");
-  CHECK(apply_input(stage, "-A") == "500ms");
-  CHECK(apply_input(stage, make_timeout_ms(500)) == "");
+  CHECK(apply_input(stage, "-A") == "+500ms");
+  CHECK(apply_input(stage, reply_timeout_ms(500)) == "");
   CHECK(apply_input(stage, "+B") == "+X");
   CHECK(apply_input(stage, "+B") == "+B");
   CHECK(apply_input(stage, "+B") == "+B");
@@ -1371,18 +1900,114 @@ TEST_CASE("Timeout Sequence #2", "[Stage]") {
   REQUIRE(stage.is_clear());
 
   CHECK(apply_input(stage, "+A") == "");
-  CHECK(apply_input(stage, "-A") == "500ms");
+  CHECK(apply_input(stage, "-A") == "+500ms");
   CHECK(apply_input(stage, "+C") == "+A -A +C");
   CHECK(apply_input(stage, "+C") == "+C");
   CHECK(apply_input(stage, "-C") == "-C");
   REQUIRE(stage.is_clear());
 
   CHECK(apply_input(stage, "+A") == "");
-  CHECK(apply_input(stage, "-A") == "500ms");
-  CHECK(apply_input(stage, make_timeout_ms(500)) == "");
+  CHECK(apply_input(stage, "-A") == "+500ms");
+  CHECK(apply_input(stage, reply_timeout_ms(500)) == "");
   CHECK(apply_input(stage, "+C") == "+A -A +C");
   CHECK(apply_input(stage, "+C") == "+C");
   CHECK(apply_input(stage, "-C") == "-C");
+  REQUIRE(stage.is_clear());
+
+  // key repeat is ignored
+  CHECK(apply_input(stage, "+E") == "+500ms");
+  CHECK(apply_input(stage, reply_timeout_ms(500)) == "");
+  CHECK(apply_input(stage, "+E") == "");
+  CHECK(apply_input(stage, "+F") == "+Y");
+  CHECK(apply_input(stage, "+F") == "+F");
+  CHECK(apply_input(stage, "+F") == "+F");
+  CHECK(apply_input(stage, "-F") == "-F -Y");
+  CHECK(apply_input(stage, "-E") == "");
+  REQUIRE(stage.is_clear());
+}
+
+//--------------------------------------------------------------------
+
+TEST_CASE("Ignore cancelled timeout", "[Stage]") {
+  // Issue #113
+  auto config = R"(
+    Q{500ms} >> C
+    (Q W) >> A
+    Q   >> B
+  )";
+  Stage stage = create_stage(config);
+
+  CHECK(apply_input(stage, "+Q") == "-500ms");
+  CHECK(apply_input(stage, reply_timeout_ms(500)) == "+C");
+  CHECK(apply_input(stage, "+Q") == "-500ms");
+  CHECK(apply_input(stage, reply_timeout_ms(500)) == "+C");
+  CHECK(apply_input(stage, "+Q") == "-500ms");
+  CHECK(apply_input(stage, reply_timeout_ms(499)) == "");
+  CHECK(apply_input(stage, "-Q") == "-C");
+  REQUIRE(stage.is_clear());
+
+  CHECK(apply_input(stage, "+Q") == "-500ms");
+  CHECK(apply_input(stage, reply_timeout_ms(499)) == "+B"); // <- unexpected
+  CHECK(apply_input(stage, "-Q") == "-B");
+  REQUIRE(stage.is_clear());
+
+  CHECK(apply_input(stage, "+Q") == "-500ms");
+  CHECK(apply_input(stage, reply_timeout_ms(499)) == "+B"); // <- unexpected
+  CHECK(apply_input(stage, "+W") == "+A");
+  CHECK(apply_input(stage, "-W") == "-A");
+  CHECK(apply_input(stage, "-Q") == "-B");
+  REQUIRE(stage.is_clear());
+
+  CHECK(apply_input(stage, "+W") == "");
+  CHECK(apply_input(stage, "-W") == "+W -W");
+  REQUIRE(stage.is_clear());
+  
+  CHECK(apply_input(stage, "+W") == "");
+  CHECK(apply_input(stage, "+Z") == "+W +Z");
+  CHECK(apply_input(stage, "-Z") == "-Z");
+  CHECK(apply_input(stage, "-W") == "-W");
+  REQUIRE(stage.is_clear());
+  
+  CHECK(apply_input(stage, "+W") == "");
+  CHECK(apply_input(stage, "+Q") == "+A");
+  CHECK(apply_input(stage, "-Q") == "");
+  CHECK(apply_input(stage, "-W") == "-A");
+
+  CHECK(apply_input(stage, "+W") == "");
+  CHECK(apply_input(stage, "+Q") == "+A");
+  CHECK(apply_input(stage, "-W") == "-A");
+  CHECK(apply_input(stage, "-Q") == "");
+  REQUIRE(stage.is_clear());
+}
+
+//--------------------------------------------------------------------
+
+TEST_CASE("Cancelled group timeout", "[Stage]") {
+  auto config = R"(
+    (A S){500ms} >> C
+    S >> R
+  )";
+  Stage stage = create_stage(config);
+
+  CHECK(apply_input(stage, "+A") == "");
+  CHECK(apply_input(stage, "+S") == "-500ms");
+  CHECK(apply_input(stage, reply_timeout_ms(500)) == "+C");
+  CHECK(apply_input(stage, "-A") == "");
+  CHECK(apply_input(stage, "-S") == "-C");
+  REQUIRE(stage.is_clear());
+
+  CHECK(apply_input(stage, "+A") == "");
+  CHECK(apply_input(stage, "+S") == "-500ms");
+  CHECK(apply_input(stage, reply_timeout_ms(499)) == "+A +R");
+  CHECK(apply_input(stage, "-A") == "-A");
+  CHECK(apply_input(stage, "-S") == "-R");
+  REQUIRE(stage.is_clear());
+
+  CHECK(apply_input(stage, "+S") == "");
+  CHECK(apply_input(stage, "+A") == "-500ms");
+  CHECK(apply_input(stage, reply_timeout_ms(499)) == "+R +A");
+  CHECK(apply_input(stage, "-A") == "-A -R"); // <- unexpected
+  CHECK(apply_input(stage, "-S") == "");
   REQUIRE(stage.is_clear());
 }
 
@@ -1391,30 +2016,57 @@ TEST_CASE("Timeout Sequence #2", "[Stage]") {
 TEST_CASE("Not Timeout Hold", "[Stage]") {
   auto config = R"(
     A{!1000ms} >> X
+    A >> Y
   )";
   Stage stage = create_stage(config);
 
-  CHECK(apply_input(stage, "+A") == "1000ms");
-  CHECK(apply_input(stage, make_timeout_ms(999)) == "");
+  CHECK(apply_input(stage, "+A") == "?1000ms");
+  CHECK(apply_input(stage, reply_timeout_ms(999)) == "");
   CHECK(apply_input(stage, "-A") == "+X -X");
   REQUIRE(stage.is_clear());
   
-  CHECK(apply_input(stage, "+A") == "1000ms");
-  CHECK(apply_input(stage, make_timeout_ms(1000)) == "+A");
-  CHECK(apply_input(stage, "+A") == "1000ms");
-  CHECK(apply_input(stage, make_timeout_ms(1000)) == "+A");
-  CHECK(apply_input(stage, "+A") == "1000ms");
+  CHECK(apply_input(stage, "+A") == "?1000ms");
+  CHECK(apply_input(stage, reply_timeout_ms(1000)) == "+Y");
+  CHECK(apply_input(stage, "+A") == "?1000ms");
+  CHECK(apply_input(stage, reply_timeout_ms(1000)) == "+Y");
+  CHECK(apply_input(stage, "+A") == "?1000ms");
   // output is suppressed when timeout was exceeded once
-  CHECK(apply_input(stage, make_timeout_ms(271)) == "");
+  CHECK(apply_input(stage, reply_timeout_ms(271)) == "");
+  CHECK(apply_input(stage, "-A") == "-Y +Y -Y");
+  REQUIRE(stage.is_clear());
+
+  CHECK(apply_input(stage, "+A") == "?1000ms");
+  CHECK(apply_input(stage, reply_timeout_ms(271)) == "");
+  CHECK(apply_input(stage, "+B") == "+Y +B");
+  CHECK(apply_input(stage, "+B") == "+B");
+  CHECK(apply_input(stage, "-B") == "-B");
+  CHECK(apply_input(stage, "-A") == "-Y");
+  REQUIRE(stage.is_clear());
+}
+
+//--------------------------------------------------------------------
+
+TEST_CASE("Not Timeout Hold with ContextActive", "[Stage]") {
+  auto config = R"(
+    ContextActive >> 
+    A{!1000ms} >> X
+  )";
+  Stage stage = create_stage(config, false);
+  REQUIRE(format_sequence(stage.set_active_client_contexts({ 0 })) == "");
+
+  CHECK(apply_input(stage, "+A") == "?1000ms");
+  CHECK(apply_input(stage, reply_timeout_ms(1000)) == "+A");
+  CHECK(apply_input(stage, "+A") == "?1000ms");
+  CHECK(apply_input(stage, reply_timeout_ms(1000)) == "+A");
+  CHECK(apply_input(stage, "+A") == "?1000ms");
+  // output is suppressed when timeout was exceeded once
+  CHECK(apply_input(stage, reply_timeout_ms(271)) == "");
   CHECK(apply_input(stage, "-A") == "-A +A -A");
   REQUIRE(stage.is_clear());
 
-  CHECK(apply_input(stage, "+A") == "1000ms");
-  CHECK(apply_input(stage, make_timeout_ms(271)) == "");
-  CHECK(apply_input(stage, "+B") == "+A +B");
-  CHECK(apply_input(stage, "+B") == "+B");
-  CHECK(apply_input(stage, "-B") == "-B");
-  CHECK(apply_input(stage, "-A") == "-A");
+  CHECK(apply_input(stage, "+A") == "?1000ms");
+  CHECK(apply_input(stage, reply_timeout_ms(999)) == "");
+  CHECK(apply_input(stage, "-A") == "+X -X");
   REQUIRE(stage.is_clear());
 }
 
@@ -1422,18 +2074,29 @@ TEST_CASE("Not Timeout Hold", "[Stage]") {
 
 TEST_CASE("Not Timeout", "[Stage]") {
   auto config = R"(
-    A !1000ms >> X
+    A !A !1000ms >> X
+    B !1000ms >> Y
   )";
   Stage stage = create_stage(config);
 
   CHECK(apply_input(stage, "+A") == "");
-  CHECK(apply_input(stage, "-A") == "1000ms");
-  CHECK(apply_input(stage, make_timeout_ms(999)) == "+X -X");
+  CHECK(apply_input(stage, "-A") == "!1000ms");
+  CHECK(apply_input(stage, reply_timeout_ms(999)) == "+X -X");
   REQUIRE(stage.is_clear());
 
   CHECK(apply_input(stage, "+A") == "");
-  CHECK(apply_input(stage, "-A") == "1000ms");
-  CHECK(apply_input(stage, make_timeout_ms(1000)) == "+A -A");
+  CHECK(apply_input(stage, "-A") == "!1000ms");
+  CHECK(apply_input(stage, reply_timeout_ms(1000)) == "+A -A");
+  REQUIRE(stage.is_clear());
+
+  CHECK(apply_input(stage, "+B") == "!1000ms");
+  CHECK(apply_input(stage, reply_timeout_ms(999)) == "+Y");
+  CHECK(apply_input(stage, "-B") == "-Y");
+  REQUIRE(stage.is_clear());
+
+  CHECK(apply_input(stage, "+B") == "!1000ms");
+  CHECK(apply_input(stage, reply_timeout_ms(1000)) == "+B");
+  CHECK(apply_input(stage, "-B") == "-B");
   REQUIRE(stage.is_clear());
 }
 
@@ -1447,36 +2110,36 @@ TEST_CASE("Timeout Xcape", "[Stage]") {
   Stage stage = create_stage(config);
 
   // long press - output Control, do not output Escape
-  CHECK(apply_input(stage, "+ControlLeft") == "500ms");
-  CHECK(apply_input(stage, make_timeout_ms(500)) == "+ControlLeft");
-  CHECK(apply_input(stage, "+ControlLeft") == "500ms");
-  CHECK(apply_input(stage, make_timeout_ms(500)) == "+ControlLeft");
-  CHECK(apply_input(stage, "+ControlLeft") == "500ms");
+  CHECK(apply_input(stage, "+ControlLeft") == "?500ms");
+  CHECK(apply_input(stage, reply_timeout_ms(500)) == "+ControlLeft");
+  CHECK(apply_input(stage, "+ControlLeft") == "?500ms");
+  CHECK(apply_input(stage, reply_timeout_ms(500)) == "+ControlLeft");
+  CHECK(apply_input(stage, "+ControlLeft") == "?500ms");
   // release while waiting for timeout - send elapsed time
-  CHECK(apply_input(stage, make_timeout_ms(271)) == "");
+  CHECK(apply_input(stage, reply_timeout_ms(271)) == "");
   // output is suppressed when timeout was exceeded once
   CHECK(apply_input(stage, "-ControlLeft") == "-ControlLeft +ControlLeft -ControlLeft");
   REQUIRE(stage.is_clear());
 
   // short press - output Escape
-  CHECK(apply_input(stage, "+ControlLeft") == "500ms");
-  CHECK(apply_input(stage, make_timeout_ms(271)) == "");
+  CHECK(apply_input(stage, "+ControlLeft") == "?500ms");
+  CHECK(apply_input(stage, reply_timeout_ms(271)) == "");
   CHECK(apply_input(stage, "-ControlLeft") == "+Escape -Escape");
   REQUIRE(stage.is_clear());
   
   // press with another key fast - output Control
-  CHECK(apply_input(stage, "+ControlLeft") == "500ms");
-  CHECK(apply_input(stage, make_timeout_ms(271)) == "");
+  CHECK(apply_input(stage, "+ControlLeft") == "?500ms");
+  CHECK(apply_input(stage, reply_timeout_ms(271)) == "");
   CHECK(apply_input(stage, "+C") == "+ControlLeft +C");
   CHECK(apply_input(stage, "-C") == "-C");
   CHECK(apply_input(stage, "-ControlLeft") == "-ControlLeft");
   REQUIRE(stage.is_clear());
 
   // press with another key slow - output Control
-  CHECK(apply_input(stage, "+ControlLeft") == "500ms");
-  CHECK(apply_input(stage, make_timeout_ms(500)) == "+ControlLeft");
-  CHECK(apply_input(stage, "+ControlLeft") == "500ms");
-  CHECK(apply_input(stage, make_timeout_ms(271)) == "");
+  CHECK(apply_input(stage, "+ControlLeft") == "?500ms");
+  CHECK(apply_input(stage, reply_timeout_ms(500)) == "+ControlLeft");
+  CHECK(apply_input(stage, "+ControlLeft") == "?500ms");
+  CHECK(apply_input(stage, reply_timeout_ms(271)) == "");
   CHECK(apply_input(stage, "+C") == "+ControlLeft +C");
   CHECK(apply_input(stage, "+C") == "+C");
   CHECK(apply_input(stage, "-C") == "-C");
@@ -1484,32 +2147,32 @@ TEST_CASE("Timeout Xcape", "[Stage]") {
   REQUIRE(stage.is_clear());
 
   // press with mouse button slow - output Control
-  CHECK(apply_input(stage, "+ControlLeft") == "500ms");
-  CHECK(apply_input(stage, make_timeout_ms(500)) == "+ControlLeft");
-  CHECK(apply_input(stage, "+ControlLeft") == "500ms");
-  CHECK(apply_input(stage, make_timeout_ms(271)) == "");
+  CHECK(apply_input(stage, "+ControlLeft") == "?500ms");
+  CHECK(apply_input(stage, reply_timeout_ms(500)) == "+ControlLeft");
+  CHECK(apply_input(stage, "+ControlLeft") == "?500ms");
+  CHECK(apply_input(stage, reply_timeout_ms(271)) == "");
   CHECK(apply_input(stage, "+ButtonLeft") == "+ControlLeft +ButtonLeft");
   CHECK(apply_input(stage, "-ButtonLeft") == "-ButtonLeft");
   // key repeat on keyboard
-  CHECK(apply_input(stage, "+ControlLeft") == "500ms");
-  CHECK(apply_input(stage, make_timeout_ms(271)) == "");
+  CHECK(apply_input(stage, "+ControlLeft") == "?500ms");
+  CHECK(apply_input(stage, reply_timeout_ms(271)) == "");
   CHECK(apply_input(stage, "-ControlLeft") == "-ControlLeft +ControlLeft -ControlLeft");
   REQUIRE(stage.is_clear());
 
   // press with mouse button fast - output Control
-  CHECK(apply_input(stage, "+ControlLeft") == "500ms");
-  CHECK(apply_input(stage, make_timeout_ms(271)) == "");
+  CHECK(apply_input(stage, "+ControlLeft") == "?500ms");
+  CHECK(apply_input(stage, reply_timeout_ms(271)) == "");
   CHECK(apply_input(stage, "+ButtonLeft") == "+ControlLeft +ButtonLeft");
   CHECK(apply_input(stage, "-ButtonLeft") == "-ButtonLeft");
   // key repeat on keyboard
-  CHECK(apply_input(stage, "+ControlLeft") == "500ms");
-  CHECK(apply_input(stage, make_timeout_ms(271)) == "");
+  CHECK(apply_input(stage, "+ControlLeft") == "?500ms");
+  CHECK(apply_input(stage, reply_timeout_ms(271)) == "");
   CHECK(apply_input(stage, "-ControlLeft") == "-ControlLeft +ControlLeft -ControlLeft");
   REQUIRE(stage.is_clear());
 
   // press with mouse button fast - output Control
-  CHECK(apply_input(stage, "+ControlLeft") == "500ms");
-  CHECK(apply_input(stage, make_timeout_ms(271)) == "");
+  CHECK(apply_input(stage, "+ControlLeft") == "?500ms");
+  CHECK(apply_input(stage, reply_timeout_ms(271)) == "");
   CHECK(apply_input(stage, "+ButtonLeft") == "+ControlLeft +ButtonLeft");
   CHECK(apply_input(stage, "-ButtonLeft") == "-ButtonLeft");
   CHECK(apply_input(stage, "-ControlLeft") == "-ControlLeft");
@@ -1525,21 +2188,21 @@ TEST_CASE("Timeout MinMax", "[Stage]") {
   Stage stage = create_stage(config);
 
   // too short
-  CHECK(apply_input(stage, "+A") == "500ms");
-  CHECK(apply_input(stage, make_timeout_ms(499)) == "+A");
+  CHECK(apply_input(stage, "+A") == "-500ms");
+  CHECK(apply_input(stage, reply_timeout_ms(499)) == "+A");
   CHECK(apply_input(stage, "-A") == "-A");
 
   // in time
-  CHECK(apply_input(stage, "+A") == "500ms");
-  CHECK(apply_input(stage, make_timeout_ms(500)) == "500ms");
-  CHECK(apply_input(stage, make_timeout_ms(499)) == "");
+  CHECK(apply_input(stage, "+A") == "-500ms");
+  CHECK(apply_input(stage, reply_timeout_ms(500)) == "?500ms");
+  CHECK(apply_input(stage, reply_timeout_ms(499)) == "");
   CHECK(apply_input(stage, "-A") == "+B -B");
 
   // too long
-  CHECK(apply_input(stage, "+A") == "500ms");
-  CHECK(apply_input(stage, make_timeout_ms(500)) == "500ms");
-  CHECK(apply_input(stage, make_timeout_ms(500)) == "+A 500ms");
-  CHECK(apply_input(stage, make_timeout_ms(271)) == "");
+  CHECK(apply_input(stage, "+A") == "-500ms");
+  CHECK(apply_input(stage, reply_timeout_ms(500)) == "?500ms");
+  CHECK(apply_input(stage, reply_timeout_ms(500)) == "+A ?500ms");
+  CHECK(apply_input(stage, reply_timeout_ms(271)) == "");
   // output is suppressed when timeout was exceeded once
   CHECK(apply_input(stage, "-A") == "-A");
 }
@@ -1551,7 +2214,7 @@ TEST_CASE("Timeout Morse", "[Stage]") {
     e = 500ms       # end
     c = !500ms      # continue
     s = X{!200ms}   # short
-    l = X{200ms}    # long
+    l = X{200ms} !X # long
 
     s c l e         >> A
     l c s c s c s e >> B
@@ -1564,63 +2227,183 @@ TEST_CASE("Timeout Morse", "[Stage]") {
   Stage stage = create_stage(config);
 
   // A
-  CHECK(apply_input(stage, "+X") == "200ms");            // waiting for short
-  CHECK(apply_input(stage, make_timeout_ms(100)) == ""); // waiting for release
-  CHECK(apply_input(stage, "-X") == "500ms");            // waiting for continue
-  CHECK(apply_input(stage, make_timeout_ms(123)) == ""); // waiting for press
-  CHECK(apply_input(stage, "+X") == "200ms");            // waiting for long
-  CHECK(apply_input(stage, make_timeout_ms(200)) == ""); // waiting for release
+  CHECK(apply_input(stage, "+X") == "?200ms");           // waiting for short
+  CHECK(apply_input(stage, reply_timeout_ms(100)) == ""); // waiting for release
+  CHECK(apply_input(stage, "-X") == "!500ms");           // waiting for continue
+  CHECK(apply_input(stage, reply_timeout_ms(123)) == ""); // waiting for press
+  CHECK(apply_input(stage, "+X") == "-200ms");           // waiting for long
+  CHECK(apply_input(stage, reply_timeout_ms(200)) == ""); // waiting for release
   CHECK(apply_input(stage, "+X") == "");                 // key repeat ignored
   CHECK(apply_input(stage, "+X") == "");                 // key repeat ignored
-  CHECK(apply_input(stage, "-X") == "500ms");            // waiting for end
-  CHECK(apply_input(stage, make_timeout_ms(500)) == "+A -A");
+  CHECK(apply_input(stage, "-X") == "+500ms");           // waiting for end
+  CHECK(apply_input(stage, reply_timeout_ms(500)) == "+A -A");
   REQUIRE(stage.is_clear());
 
   // ERROR
-  CHECK(apply_input(stage, "+X") == "200ms");            
-  CHECK(apply_input(stage, make_timeout_ms(200)) == ""); // long
+  CHECK(apply_input(stage, "+X") == "?200ms");            
+  CHECK(apply_input(stage, reply_timeout_ms(200)) == ""); // long
   CHECK(apply_input(stage, "+X") == "");                 // key repeat ignored
   CHECK(apply_input(stage, "+X") == "");                 // key repeat ignored
-  CHECK(apply_input(stage, "-X") == "500ms");
-  CHECK(apply_input(stage, make_timeout_ms(500)) == "+Y -Y");
+  CHECK(apply_input(stage, "-X") == "!500ms");
+  CHECK(apply_input(stage, reply_timeout_ms(500)) == "+Y -Y");
   REQUIRE(stage.is_clear());
 
   // E
-  CHECK(apply_input(stage, "+X") == "200ms");            // waiting for short
-  CHECK(apply_input(stage, make_timeout_ms(100)) == "");
-  CHECK(apply_input(stage, "-X") == "500ms");            // waiting for end
-  CHECK(apply_input(stage, make_timeout_ms(500)) == "+E -E");
+  CHECK(apply_input(stage, "+X") == "?200ms");           // waiting for short
+  CHECK(apply_input(stage, reply_timeout_ms(100)) == "");
+  CHECK(apply_input(stage, "-X") == "!500ms");           // waiting for end
+  CHECK(apply_input(stage, reply_timeout_ms(500)) == "+E -E");
   REQUIRE(stage.is_clear());
 
   // ERROR
-  CHECK(apply_input(stage, "+X") == "200ms");
-  CHECK(apply_input(stage, make_timeout_ms(200)) == ""); // long
+  CHECK(apply_input(stage, "+X") == "?200ms");
+  CHECK(apply_input(stage, reply_timeout_ms(200)) == ""); // long
   CHECK(apply_input(stage, "+X") == "");                 // key repeat ignored
   CHECK(apply_input(stage, "+X") == "");                 // key repeat ignored
-  CHECK(apply_input(stage, "-X") == "500ms");
-  CHECK(apply_input(stage, make_timeout_ms(123)) == "");
-  CHECK(apply_input(stage, "+X") == "200ms");
-  CHECK(apply_input(stage, make_timeout_ms(200)) == "+Y -Y"); // long (restart)
+  CHECK(apply_input(stage, "-X") == "!500ms");
+  CHECK(apply_input(stage, reply_timeout_ms(123)) == "");
+  CHECK(apply_input(stage, "+X") == "?200ms");
+  CHECK(apply_input(stage, reply_timeout_ms(200)) == "+Y -Y"); // long (restart)
   CHECK(apply_input(stage, "+X") == "");                 // key repeat ignored
   CHECK(apply_input(stage, "+X") == "");                 // key repeat ignored
-  CHECK(apply_input(stage, "-X") == "500ms");
-  CHECK(apply_input(stage, make_timeout_ms(500)) == "+Y -Y");
+  CHECK(apply_input(stage, "-X") == "!500ms");
+  CHECK(apply_input(stage, reply_timeout_ms(500)) == "+Y -Y");
   REQUIRE(stage.is_clear());
 
   // D
-  CHECK(apply_input(stage, "+X") == "200ms");            // waiting for long
-  CHECK(apply_input(stage, make_timeout_ms(200)) == ""); // waiting for release
+  CHECK(apply_input(stage, "+X") == "?200ms");           // waiting for long
+  CHECK(apply_input(stage, reply_timeout_ms(200)) == ""); // waiting for release
   CHECK(apply_input(stage, "+X") == "");                 // key repeat ignored
   CHECK(apply_input(stage, "+X") == "");                 // key repeat ignored
-  CHECK(apply_input(stage, "-X") == "500ms");            // waiting for continue
-  CHECK(apply_input(stage, make_timeout_ms(123)) == ""); // waiting for press
-  CHECK(apply_input(stage, "+X") == "200ms");            // waiting for short
-  CHECK(apply_input(stage, make_timeout_ms(100)) == ""); // waiting for release
-  CHECK(apply_input(stage, "-X") == "500ms");            // waiting for continue
-  CHECK(apply_input(stage, make_timeout_ms(123)) == ""); // waiting for press
-  CHECK(apply_input(stage, "+X") == "200ms");            // waiting for short
-  CHECK(apply_input(stage, make_timeout_ms(100)) == ""); // waiting for release
-  CHECK(apply_input(stage, "-X") == "500ms");            // waiting for end
-  CHECK(apply_input(stage, make_timeout_ms(500)) == "+D -D");
+  CHECK(apply_input(stage, "-X") == "!500ms");           // waiting for continue
+  CHECK(apply_input(stage, reply_timeout_ms(123)) == ""); // waiting for press
+  CHECK(apply_input(stage, "+X") == "?200ms");           // waiting for short
+  CHECK(apply_input(stage, reply_timeout_ms(100)) == ""); // waiting for release
+  CHECK(apply_input(stage, "-X") == "!500ms");           // waiting for continue
+  CHECK(apply_input(stage, reply_timeout_ms(123)) == ""); // waiting for press
+  CHECK(apply_input(stage, "+X") == "?200ms");           // waiting for short
+  CHECK(apply_input(stage, reply_timeout_ms(100)) == ""); // waiting for release
+  CHECK(apply_input(stage, "-X") == "!500ms");           // waiting for end
+  CHECK(apply_input(stage, reply_timeout_ms(500)) == "+D -D");
+  REQUIRE(stage.is_clear());
+}
+
+//--------------------------------------------------------------------
+
+TEST_CASE("Output Timeout", "[Stage]") {
+  auto config = R"(
+    A >> X 1000ms Y
+    B >> Z 1000ms
+  )";
+  Stage stage = create_stage(config);
+
+  CHECK(apply_input(stage, "+A") == "+X -X +1000ms +Y -Y");
+  CHECK(apply_input(stage, "-A") == "");
+
+  CHECK(apply_input(stage, "+B") == "+Z -Z +1000ms");
+  CHECK(apply_input(stage, "-B") == "");
+}
+
+//--------------------------------------------------------------------
+
+TEST_CASE("Explicit modifier", "[Stage]") {
+  auto config = R"(
+    Shift >> Shift
+    A >> !Shift E Shift{F}                # eF
+    B >> !Shift E Shift{F} E Shift{F}     # eFeF
+    C >> !Shift Shift{G} H                # Gh
+    D >> !Shift Shift{G} H Shift{G} H     # GhGh
+  )";
+  Stage stage = create_stage(config);
+
+  REQUIRE(apply_input(stage, "+ShiftLeft") == "+ShiftLeft");
+  REQUIRE(apply_input(stage, "+A") == "-ShiftLeft +E -E +ShiftLeft +F -F -ShiftLeft");
+  REQUIRE(apply_input(stage, "-A") == "");
+  REQUIRE(apply_input(stage, "+A") == "+E -E +ShiftLeft +F -F -ShiftLeft");
+  REQUIRE(apply_input(stage, "-A") == "");
+  // that the Shift is released at the end might be suboptimal, at least it is restored
+  REQUIRE(apply_input(stage, "+R") == "+ShiftLeft +R");
+  REQUIRE(apply_input(stage, "-R") == "-R");
+  REQUIRE(apply_input(stage, "-ShiftLeft") == "-ShiftLeft");
+  REQUIRE(stage.is_clear());
+
+  REQUIRE(apply_input(stage, "+ShiftLeft") == "+ShiftLeft");
+  REQUIRE(apply_input(stage, "+B") == 
+    "-ShiftLeft +E -E +ShiftLeft +F -F -ShiftLeft +E -E +ShiftLeft +F -F -ShiftLeft");
+  REQUIRE(apply_input(stage, "-B") == "");
+  REQUIRE(apply_input(stage, "+B") == 
+    "+E -E +ShiftLeft +F -F -ShiftLeft +E -E +ShiftLeft +F -F -ShiftLeft");
+  REQUIRE(apply_input(stage, "-B") == "");
+  REQUIRE(apply_input(stage, "-ShiftLeft") == "");
+  REQUIRE(stage.is_clear());
+
+  REQUIRE(apply_input(stage, "+ShiftLeft") == "+ShiftLeft");
+  // that the Shift is released and pressed at the beginning is also not ideal
+  REQUIRE(apply_input(stage, "+C") == "-ShiftLeft +ShiftLeft +G -G -ShiftLeft +H -H");
+  REQUIRE(apply_input(stage, "-C") == "");
+  REQUIRE(apply_input(stage, "+C") == "+ShiftLeft +G -G -ShiftLeft +H -H");
+  REQUIRE(apply_input(stage, "-C") == "");
+  REQUIRE(apply_input(stage, "+R") == "+ShiftLeft +R");
+  REQUIRE(apply_input(stage, "-R") == "-R");
+  REQUIRE(apply_input(stage, "-ShiftLeft") == "-ShiftLeft");
+  REQUIRE(stage.is_clear());
+
+  REQUIRE(apply_input(stage, "+ShiftLeft") == "+ShiftLeft");
+  REQUIRE(apply_input(stage, "+D") == 
+    "-ShiftLeft +ShiftLeft +G -G -ShiftLeft +H -H +ShiftLeft +G -G -ShiftLeft +H -H");
+  REQUIRE(apply_input(stage, "-D") == "");
+  REQUIRE(apply_input(stage, "+D") == 
+    "+ShiftLeft +G -G -ShiftLeft +H -H +ShiftLeft +G -G -ShiftLeft +H -H");
+  REQUIRE(apply_input(stage, "-D") == "");
+  REQUIRE(apply_input(stage, "-ShiftLeft") == "");
+  REQUIRE(stage.is_clear());
+}
+
+//--------------------------------------------------------------------
+
+TEST_CASE("String typing", "[Stage]") {
+  auto config = R"(
+    Shift >> Shift
+    A >> "eF"
+    B >> "eFeF"
+    C >> "Gh"
+    D >> "GhGh"
+  )";
+  Stage stage = create_stage(config);
+
+  REQUIRE(apply_input(stage, "+ShiftLeft") == "+ShiftLeft");
+  REQUIRE(apply_input(stage, "+A") == "-ShiftLeft +E -E +ShiftLeft +F -F -ShiftLeft");
+  REQUIRE(apply_input(stage, "-A") == "");
+  REQUIRE(apply_input(stage, "+A") == "+E -E +ShiftLeft +F -F -ShiftLeft");
+  REQUIRE(apply_input(stage, "-A") == "");
+  REQUIRE(apply_input(stage, "-ShiftLeft") == "");
+  REQUIRE(stage.is_clear());
+
+  REQUIRE(apply_input(stage, "+ShiftLeft") == "+ShiftLeft");
+  REQUIRE(apply_input(stage, "+B") == 
+    "-ShiftLeft +E -E +ShiftLeft +F -F -ShiftLeft +E -E +ShiftLeft +F -F -ShiftLeft");
+  REQUIRE(apply_input(stage, "-B") == "");
+  REQUIRE(apply_input(stage, "+B") == 
+    "+E -E +ShiftLeft +F -F -ShiftLeft +E -E +ShiftLeft +F -F -ShiftLeft");
+  REQUIRE(apply_input(stage, "-B") == "");
+  REQUIRE(apply_input(stage, "-ShiftLeft") == "");
+  REQUIRE(stage.is_clear());
+
+  REQUIRE(apply_input(stage, "+ShiftLeft") == "+ShiftLeft");
+  REQUIRE(apply_input(stage, "+C") == "+G -G -ShiftLeft +H -H");
+  REQUIRE(apply_input(stage, "-C") == "");
+  REQUIRE(apply_input(stage, "+C") == "+ShiftLeft +G -G -ShiftLeft +H -H");
+  REQUIRE(apply_input(stage, "-C") == "");
+  REQUIRE(apply_input(stage, "-ShiftLeft") == "");
+  REQUIRE(stage.is_clear());
+
+  REQUIRE(apply_input(stage, "+ShiftLeft") == "+ShiftLeft");
+  REQUIRE(apply_input(stage, "+D") == 
+    "+G -G -ShiftLeft +H -H +ShiftLeft +G -G -ShiftLeft +H -H");
+  REQUIRE(apply_input(stage, "-D") == "");
+  REQUIRE(apply_input(stage, "+D") == 
+    "+ShiftLeft +G -G -ShiftLeft +H -H +ShiftLeft +G -G -ShiftLeft +H -H");
+  REQUIRE(apply_input(stage, "-D") == "");
+  REQUIRE(apply_input(stage, "-ShiftLeft") == "");
   REQUIRE(stage.is_clear());
 }
